ArgParse

Description

This module handles parsing CLI arguments into Commands. See the excellent documentation at https://hackage.haskell.org/package/optparse-applicative

Documentation

data RunSource

Valid ways to provide source code to the run command

Constructors

RunFromPipe (HashQualified Name)

RunFromSymbol ProjectPathNames

RunFromFile FilePath (HashQualified Name)

RunCompiled FilePath

data ShouldForkCodebase

Constructors

UseFork

DontFork

data ShouldSaveCodebase

Constructors

SaveCodebase (Maybe FilePath)

DontSaveCodebase

data CodebasePathOption

Constructors

CreateCodebaseWhenMissing FilePath

DontCreateCodebaseWhenMissing FilePath

data ShouldExit

Constructors

Exit

DoNotExit

data IsHeadless

Constructors

Headless

WithCLI

data Command

Represents commands the cli can run.

Note that this is not one-to-one with command-parsers since some are simple variants. E.g. run, run.file, run.pipe

Constructors

Launch IsHeadless CodebaseServerOpts (Maybe (ProjectAndBranch ProjectName ProjectBranchName)) ShouldWatchFiles

PrintVersion

Init

Run RunSource [String]

Transcript ShouldForkCodebase ShouldSaveCodebase (Maybe RtsStatsPath) (NonEmpty FilePath)

data GlobalOptions

Options shared by sufficiently many subcommands.

Constructors

GlobalOptions

Fields

codebasePathOption :: Maybe CodebasePathOption

exitOption :: ShouldExit

nativeRuntimePath :: Maybe FilePath

lspFormattingConfig :: LspFormattingConfig

rootParserInfo :: String -> String -> CodebaseServerOpts -> ParserInfo (GlobalOptions, Command)

The root-level ParserInfo.

type UsageRenderer

Arguments

= Maybe String
Optional sub-command to render help for

-> String

parseCLIArgs :: String -> String -> IO (UsageRenderer, GlobalOptions, Command)

Parse the command description, options, and usage information from provided cli arguments.

codebaseServerOptsFromEnv :: IO CodebaseServerOpts

Load default options from environment variables.

renderUsage :: String -> ParserInfo a -> ParserPrefs -> Maybe String -> String

Purely renders the full help summary for the CLI, or an optional subcommand.

versionCommand :: Mod CommandFields Command

initCommand :: Mod CommandFields Command

runDesc :: String -> String -> String

runSymbolCommand :: Mod CommandFields Command

runFileCommand :: Mod CommandFields Command

runPipeCommand :: Mod CommandFields Command

runCompiledCommand :: Mod CommandFields Command

transcriptCommand :: Mod CommandFields Command

transcriptForkCommand :: Mod CommandFields Command

commandParser :: CodebaseServerOpts -> Parser Command

globalOptionsParser :: Parser GlobalOptions

codebasePathParser :: Parser (Maybe CodebasePathOption)

codebaseCreateParser :: Parser (Maybe CodebasePathOption)

exitParser :: Parser ShouldExit

lspFormattingParser :: Parser LspFormattingConfig

versionOptionParser :: String -> String -> Parser (a -> a)

launchHeadlessCommand :: CodebaseServerOpts -> Mod CommandFields Command

codebaseServerOptsParser :: CodebaseServerOpts -> Parser CodebaseServerOpts

launchParser :: CodebaseServerOpts -> IsHeadless -> Parser Command

initParser :: Parser Command

versionParser :: Parser Command

runArgumentParser :: Parser [String]

runHQParser :: Parser (HashQualified Name)

runProjectPathParser :: Parser ProjectPathNames

runSymbolParser :: Parser Command

runFileParser :: Parser Command

runPipeParser :: Parser Command

runCompiledParser :: Parser Command

rtsStatsOption :: Parser (Maybe RtsStatsPath)

saveCodebaseFlag :: Parser ShouldSaveCodebase

saveCodebaseToFlag :: Parser ShouldSaveCodebase

startingProjectOption :: Parser (Maybe (ProjectAndBranch ProjectName ProjectBranchName))

noFileWatchFlag :: Parser ShouldWatchFiles

readAbsolutePath :: ReadM Absolute

nativeRuntimePathFlag :: Parser (Maybe FilePath)

readPath' :: ReadM Path'

readProjectAndBranchNames :: ReadM (ProjectAndBranch ProjectName ProjectBranchName)

fileArgument :: String -> Parser FilePath

transcriptParser :: Parser Command

transcriptForkParser :: Parser Command

unisonHelp :: String -> String -> Doc