Unison.Auth.CredentialManager

Documentation

saveCredentials :: MonadUnliftIO m => CredentialManager -> CodeserverId -> CodeserverCredentials -> m ()

Saves credentials to the active profile.

data CredentialManager

A CredentialManager knows how to load, save, and cache credentials. It's thread-safe and safe for use across multiple UCM clients. Note: Currently the in-memory cache is _not_ updated if a different UCM updates the credentials file, however this shouldn't pose any problems, since auth will still be refreshed if we encounter any auth failures on requests.

newCredentialManager :: MonadIO m => m CredentialManager

getCredentials :: MonadIO m => CredentialManager -> CodeserverId -> m (Either CredentialFailure CodeserverCredentials)

isExpired :: MonadIO m => CodeserverCredentials -> m Bool

Checks whether CodeserverCredentials are expired.