Unison.Auth.HTTPClient

Documentation

newAuthorizedHTTPClient :: MonadIO m => CredentialManager -> UCMVersion -> m Manager

Returns a new http manager which applies the appropriate Authorization header to any hosts our UCM is authenticated with.

authMiddleware :: TokenProvider -> Request -> IO Request

Adds Bearer tokens to requests according to their host. If a CredentialFailure occurs (failure to refresh a token), auth is simply omitted, and the request is likely to trigger a 401 response which the caller can detect and initiate a re-auth.

If a host isn't associated with any credentials auth is omitted.