Unison.Auth.Tokens

Documentation

type TokenProvider = CodeserverId -> IO (Either CredentialFailure AccessToken)

Given a CodeserverId, provide a valid AccessToken for the associated host. The TokenProvider may automatically refresh access tokens if we have a refresh token.

accessTokenEnvVarKey :: String

If provided, this access token will be used on all requests which use the Authenticated HTTP Client; i.e. all codeserver interactions.

It's useful in scripted contexts or when running transcripts against a codeserver.

newTokenProvider :: CredentialManager -> TokenProvider

Creates a TokenProvider using the given CredentialManager

performTokenRefresh :: MonadIO m => DiscoveryDoc -> Tokens -> m (Either CredentialFailure Tokens)

Don't yet support automatically refreshing tokens.

Specification: https://datatracker.ietf.org/doc/html/rfc6749#section-6