Unison.Cli.DownloadUtils

Description

Utility functions for downloading remote entities and storing them locally in SQLite.

These are shared by commands like pull and clone.

Documentation

downloadProjectBranchFromShare

Arguments

:: HasCallStack

=> IncludeSquashedHead

-> RemoteProjectBranch

-> Bool
Whether this download is part of a pull operation. If True, we will show the from and to causal hashes.

-> Cli (Either ShareError CausalHash)

Download a project/branch from Share.

downloadLooseCodeFromShare :: ReadShareLooseCode -> Cli (Either ShareError CausalHash)

Download loose code from Share.

data SyncVersion

Constructors

SyncV1

SyncV2