Unison.Cli.Monad

Description

The main CLI monad.

Cli monad

data Cli a

The main command-line app monad.

It is a reader monad of Env.

It is a state monad of LoopState.

It is a short-circuiting monad: a Cli computation can short-circuit with success or failure in a delimited scope.

It is a resource monad: resources can be acquired in callback-style.

It is an IO monad: you can do IO things, but throwing synchronous exceptions is discouraged. Use the built-in short-circuiting mechanism instead.

data ReturnType a

What a Cli action returns: a value, an instruction to continue processing input, or an instruction to stop processing input.

Constructors

Success a

Continue

HaltRepl

type SourceName = Text

Name used for a source-file/source buffer

runCli :: Env -> LoopState -> Cli a -> IO (ReturnType a, LoopState)

Run a Cli action down to IO.

Envronment

data Env

The command-line app monad environment.

Get the environment with ask.

Constructors

Env

Fields

authHTTPClient :: AuthenticatedHttpClient

codebase :: Codebase IO Symbol Ann

credentialManager :: CredentialManager

generateUniqueName :: IO UniqueName

Generate a unique name.

loadSource :: SourceName -> IO LoadSourceResult

How to load source code.

writeSource :: SourceName -> Text -> Bool -> IO ()

How to write source code. Bool = make new fold?

notify :: Output -> IO ()

What to do with output for the user.

notifyNumbered :: NumberedOutput -> IO NumberedArgs

What to do with numbered output for the user.

runtime :: Runtime Symbol

sandboxedRuntime :: Runtime Symbol

nativeRuntime :: Runtime Symbol

serverBaseUrl :: Maybe BaseUrl

ucmVersion :: UCMVersion

isTranscriptTest :: Bool

Whether we're running in a transcript test or not. Avoid using this except when absolutely necessary.

Immutable state

data LoopState

The command-line app monad mutable state.

There's an additional pseudo "currentPath" field lens, for convenience.

Constructors

LoopState

Fields

projectPathStack :: NonEmpty ProjectPathIds

latestFile :: Maybe (FilePath, Bool)

latestTypecheckedFile :: Maybe (Either (UnisonFile Symbol Ann) (TypecheckedUnisonFile Symbol Ann))

lastInput :: Maybe Input

numberedArgs :: NumberedArgs

lastRunResult :: Maybe (Term Symbol Ann, Type Symbol Ann, TypecheckedUnisonFile Symbol Ann)

loopState0 :: ProjectPathIds -> LoopState

Create an initial loop state given a root branch and the current path.

getProjectPathIds :: Cli ProjectPathIds

Lifting IO actions

ioE :: IO (Either e a) -> (e -> Cli a) -> Cli a

Lift an action of type IO (Either e a), given a continuation for e.

Acquiring resources

with :: (forall x. (a -> IO x) -> IO x) -> (a -> Cli b) -> Cli b

Wrap a continuation with Cli.

Useful for resource acquisition:

with (bracket create destroy) \resource ->
  ...

with_ :: (forall x. IO x -> IO x) -> Cli a -> Cli a

A variant of with for actions that don't acquire a resource (like bracket_).

withE :: (forall x. (a -> IO x) -> IO (Either e x)) -> (Either e a -> Cli b) -> Cli b

A variant of with for the variant of bracketing function that may return a Left rather than call the provided continuation.

Short-circuiting

label :: forall a. ((forall void. a -> Cli void) -> Cli a) -> Cli a

Create a label that can be jumped to.

x <- label \j0 -> do
  ...
  label \j1 -> do
    ...
    j0 someValue
    ... -- We don't get here
  ... -- We don't get here
-- x is bound to someValue

labelE :: ((forall void. a -> Cli void) -> Cli b) -> Cli (Either a b)

A variant of label for the common case that early-return values are tagged with a Left.

returnEarly :: Output -> Cli a

Short-circuit the processing of the current input.

returnEarlyWithoutOutput :: Cli a

Variant of returnEarly that doesn't take a final output message.

haltRepl :: Cli a

Stop processing inputs from the user.

Changing the current directory

cd :: Absolute -> Cli ()

popd :: Cli Bool

Pop the latest path off the stack, if it's not the only path in the stack.

Returns whether anything was popped.

switchProject :: ProjectAndBranch ProjectId ProjectBranchId -> Cli ()

Communicating output to the user

respond :: Output -> Cli ()

respondNumbered :: NumberedOutput -> Cli ()

setNumberedArgs :: NumberedArgs -> Cli ()

Updates the numbered args, but only if the new args are non-empty.

Debug-timing actions

time :: String -> Cli a -> Cli a

Time an action.

Running transactions

runTransaction :: Transaction a -> Cli a

runTransactionWithRollback :: ((forall void. Output -> Transaction void) -> Transaction a) -> Cli a

Run a transaction that can abort early with an output message. todo: rename to runTransactionWithReturnEarly

runTransactionWithRollback2 :: ((forall void. a -> Transaction void) -> Transaction a) -> Cli a

Run a transaction that can abort early. todo: rename to runTransactionWithRollback

Internal

setMostRecentProjectPath :: ProjectPathIds -> Cli ()

Misc types

data LoadSourceResult

The result of calling $sel:loadSource:Env.

Constructors

InvalidSourceNameError

LoadError

LoadSuccess Text