Unison.Cli.NamesUtils

Description

Utilities that have to do with constructing names objects.

Documentation

basicParseNames :: Cli Names

basicPrettyPrintNamesA :: Cli Names

displayNames :: Var v => TypecheckedUnisonFile v a -> Cli NamesWithHistory

Produce a Names needed to display all the hashes used in the given file.

findHistoricalHQs :: Set (HashQualified Name) -> Cli Names

getBasicPrettyPrintNames :: Cli Names

makeHistoricalParsingNames :: Set (HashQualified Name) -> Cli NamesWithHistory

makePrintNamesFromLabeled' :: Set LabeledDependency -> Cli NamesWithHistory

makeShadowedPrintNamesFromHQ :: Set (HashQualified Name) -> Names -> Cli NamesWithHistory