Unison.Cli.Pretty

Description

Small combinators that pretty-print small types in a canonical way for human consumption, such as hashes, file paths, and project names.

Documentation

displayBranchHash :: CausalHash -> Text

Displays a full, non-truncated Branch.CausalHash to a string, e.g. #abcdef

prettyAbsolute :: Absolute -> Pretty

prettyProjectPath :: ProjectPath -> Pretty

prettyBranchRelativePath :: BranchRelativePath -> Pretty

prettyBase32Hex# :: IsString s => Base32Hex -> Pretty s

prettyBase32Hex :: IsString s => Base32Hex -> Pretty s

prettyBranchId :: AbsBranchId -> Pretty

prettyCausalHash :: IsString s => CausalHash -> Pretty s

prettyDeclPair :: Var v => PrettyPrintEnv -> (Reference, DisplayObject () (Decl v a)) -> Pretty

prettyDeclTriple :: Var v => (HashQualified Name, Reference, DisplayObject () (Decl v a)) -> Pretty

prettyFilePath :: FilePath -> Pretty

prettyHash :: IsString s => Hash -> Pretty s

prettyHash32 :: IsString s => Hash32 -> Pretty s

prettyHumanReadableTime :: UTCTime -> UTCTime -> Pretty

prettyLabeledDependencies :: PrettyPrintEnv -> Set LabeledDependency -> Pretty

prettyPath :: Path -> Pretty

prettyPath' :: Path' -> Pretty

prettyMergeSource :: MergeSource -> Pretty

prettyMergeSourceOrTarget :: MergeSourceOrTarget -> Pretty

prettyProjectAndBranchName :: ProjectAndBranch ProjectName ProjectBranchName -> Pretty

prettyProjectBranchName :: ProjectBranchName -> Pretty

prettyProjectName :: ProjectName -> Pretty

prettyProjectNameSlash :: ProjectName -> Pretty

prettyProjectName with a trailing slash.

prettyNamespaceKey :: Either ProjectPath (ProjectAndBranch Project ProjectBranch) -> Pretty

prettyReadRemoteNamespace :: ReadRemoteNamespace RemoteProjectBranch -> Pretty

prettyReadRemoteNamespaceWith :: (a -> Text) -> ReadRemoteNamespace a -> Pretty

prettyRelative :: Relative -> Pretty

prettyRemoteBranchInfo :: (URI, ProjectName, ProjectBranchName) -> Pretty

prettyRepoInfo :: RepoInfo -> Pretty

prettySCH :: IsString s => ShortCausalHash -> Pretty s

prettySemver :: Semver -> Pretty

prettySharePath :: Path -> Pretty

prettyShareURI :: URI -> Pretty

prettySlashProjectBranchName :: ProjectBranchName -> Pretty

Like prettyProjectBranchName, but with a leading forward slash. This is used in some outputs to encourage/advertise an unambiguous syntax for project branches, as there's an ambiguity with single-segment relative paths.

Not all project branches are printed such: for example, when listing all branches of a project, we probably don't need or want to prefix every one with a forward slash.

prettyTerm :: PrettyPrintEnvDecl -> Bool -> Bool -> (HashQualified Name, Reference, DisplayObject (Type Symbol Ann) (Term Symbol Ann)) -> Pretty SyntaxText

prettyTermName :: PrettyPrintEnv -> Referent -> Pretty

prettyType :: PrettyPrintEnvDecl -> (HashQualified Name, Reference, DisplayObject () (Decl Symbol Ann)) -> Pretty SyntaxText

prettyTypeName :: PrettyPrintEnv -> Reference -> Pretty

prettyTypeResultHeader' :: Var v => TypeResult' v a -> Pretty

prettyTypeResultHeaderFull' :: Var v => TypeResult' v a -> Pretty

prettyURI :: URI -> Pretty

prettyUnisonFile :: forall v a. (Var v, Ord a) => PrettyPrintEnvDecl -> UnisonFile v a -> Pretty ColorText

prettyWhichBranchEmpty :: WhichBranchEmpty -> Pretty

Pretty-print a WhichBranchEmpty.

prettyWriteRemoteNamespace :: ProjectAndBranch ProjectName ProjectBranchName -> Pretty

shareOrigin :: Text

unsafePrettyTermResultSigFull' :: Var v => PrettyPrintEnv -> TermResult' v a -> Pretty