Unison.Cli.ProjectUtils

Description

Project-related utilities.

Project/path helpers

expectProjectBranchByName :: Project -> ProjectBranchName -> Cli ProjectBranch

resolveBranchRelativePath :: BranchRelativePath -> Cli ProjectPath

resolveProjectBranch :: ProjectAndBranch (Maybe ProjectName) (Maybe ProjectBranchName) -> Cli (ProjectAndBranch Project ProjectBranch)

Expect/resolve branch reference with the following rules:

If the project is missing, use the current project.

If we have an unambiguous `branch` or `projectbranch`, resolve it using the current project, defaulting to main if branch is unspecified.

resolveProjectBranchInProject :: Project -> ProjectAndBranch (Maybe ProjectName) (Maybe ProjectBranchName) -> Cli (ProjectAndBranch Project ProjectBranch)

Expect/resolve branch reference with the following rules:

If the project is missing, use the provided project.

If we have an unambiguous `branch` or `projectbranch`, resolve it using the provided project, defaulting to main if branch is unspecified.

Name hydration

hydrateNames :: These ProjectName ProjectBranchName -> Cli (ProjectAndBranch ProjectName ProjectBranchName)

Loading local project info

expectProjectAndBranchByIds :: ProjectAndBranch ProjectId ProjectBranchId -> Transaction (ProjectAndBranch Project ProjectBranch)

getProjectAndBranchByTheseNames :: These ProjectName ProjectBranchName -> Cli (Maybe (ProjectAndBranch Project ProjectBranch))

getProjectAndBranchByNames :: ProjectAndBranch ProjectName ProjectBranchName -> Transaction (Maybe (ProjectAndBranch Project ProjectBranch))

getProjectByName :: ProjectName -> Cli (Maybe Project)

expectProjectAndBranchByTheseNames :: These ProjectName ProjectBranchName -> Cli (ProjectAndBranch Project ProjectBranch)

expectProjectAndBranchByTheseNamesTx :: (forall void. Output -> Transaction void) -> Project -> These ProjectName ProjectBranchName -> Transaction (ProjectAndBranch Project ProjectBranch)

Like expectProjectAndBranchByTheseNames, but in Transaction, and takes the current project and a rollback function as arguments.

getProjectBranchCausalHash :: ProjectBranch -> Transaction CausalHash

Get the causal hash of a project branch.

Loading remote project info

expectRemoteProjectById :: RemoteProjectId -> ProjectName -> Cli RemoteProject

Expect a remote project by id. Its latest-known name is also provided, for error messages.

expectRemoteProjectByName :: ProjectName -> Cli RemoteProject

expectRemoteProjectBranchById :: IncludeSquashedHead -> ProjectAndBranch (RemoteProjectId, ProjectName) (RemoteProjectBranchId, ProjectBranchName) -> Cli RemoteProjectBranch

loadRemoteProjectBranchByName :: IncludeSquashedHead -> ProjectAndBranch RemoteProjectId ProjectBranchName -> Cli (Maybe RemoteProjectBranch)

expectRemoteProjectBranchByName :: IncludeSquashedHead -> ProjectAndBranch (RemoteProjectId, ProjectName) ProjectBranchName -> Cli RemoteProjectBranch

loadRemoteProjectBranchByNames :: IncludeSquashedHead -> ProjectAndBranch ProjectName ProjectBranchName -> Cli (Maybe RemoteProjectBranch)

expectRemoteProjectBranchByNames :: IncludeSquashedHead -> ProjectAndBranch ProjectName ProjectBranchName -> Cli RemoteProjectBranch

expectRemoteProjectBranchByTheseNames :: IncludeSquashedHead -> These ProjectName ProjectBranchName -> Cli RemoteProjectBranch

Projecting out common things

justTheIds :: ProjectAndBranch Project ProjectBranch -> ProjectAndBranch ProjectId ProjectBranchId

justTheIds' :: ProjectBranch -> ProjectAndBranch ProjectId ProjectBranchId

justTheNames :: ProjectAndBranch Project ProjectBranch -> ProjectAndBranch ProjectName ProjectBranchName

Other helpers

findTemporaryBranchName :: ProjectId -> ProjectBranchName -> Transaction ProjectBranchName

expectLatestReleaseBranchName :: RemoteProject -> Cli ProjectBranchName

Expect the given remote project to have a latest release, and return it as a valid branch name.

Export fields so we can use dot-notation

data ProjectBranch

A project branch.

Constructors

ProjectBranch

Fields

projectId :: !ProjectId

branchId :: !ProjectBranchId

name :: !ProjectBranchName

parentBranchId :: !(Maybe ProjectBranchId)

isMerge :: !Bool

isUpdate :: !Bool

isUpgrade :: !Bool

data Project

A project.

Constructors

Project

Fields

projectId :: !ProjectId

name :: !ProjectName