Unison.Cli.Share.Projects

Description

This module contains Share API calls related to projects, wrapped in the Cli monad.

Here, we also validate inputs from Share that the API itself does not. For example, in the API, a project name is just a Text. But because our client requires a richer structure for project names, we try parsing them into a ProjectName, and fail right away if parsing fails.

API types

data RemoteProject

A remote project.

Constructors

RemoteProject

Fields

projectId :: RemoteProjectId

projectName :: ProjectName

latestRelease :: Maybe Semver

data RemoteProjectBranch

A remote project branch.

Constructors

RemoteProjectBranch

Fields

projectId :: RemoteProjectId

projectName :: ProjectName

branchId :: RemoteProjectBranchId

branchName :: ProjectBranchName

branchHead :: HashJWT

squashedBranchHead :: Maybe HashJWT

API functions

getProjectById :: RemoteProjectId -> Cli (Maybe RemoteProject)

Get a project by id.

On success, update the remote_project table.

getProjectByName :: ProjectName -> Cli (Maybe RemoteProject)

Get a project by name.

On success, update the remote_project table.

getProjectByName' :: ProjectName -> Cli (Either ClientError (Maybe RemoteProject))

Variant of getProjectByName that returns servant client errors.

createProject :: ProjectName -> Cli (Maybe RemoteProject)

Create a new project. Kinda weird: returns Nothing if the user handle part of the project doesn't exist.

On success, update the remote_project table.

data GetProjectBranchResponse

Constructors

GetProjectBranchResponseBranchNotFound

GetProjectBranchResponseProjectNotFound

GetProjectBranchResponseSuccess !RemoteProjectBranch

data IncludeSquashedHead

Constructors

IncludeSquashedHead

NoSquashedHead

getProjectBranchById :: IncludeSquashedHead -> ProjectAndBranch RemoteProjectId RemoteProjectBranchId -> Cli GetProjectBranchResponse

Get a project branch by id.

On success, update the remote_project_branch table.

getProjectBranchByName :: IncludeSquashedHead -> ProjectAndBranch RemoteProjectId ProjectBranchName -> Cli GetProjectBranchResponse

Get a project branch by name.

On success, update the remote_project_branch table.

getProjectBranchByName' :: IncludeSquashedHead -> ProjectAndBranch RemoteProjectId ProjectBranchName -> Cli (Either ClientError GetProjectBranchResponse)

Variant of getProjectBranchByName that returns servant client errors.

createProjectBranch :: CreateProjectBranchRequest -> Cli (Maybe RemoteProjectBranch)

Create a new project branch.

On success, update the remote_project_branch table.

data SetProjectBranchHeadResponse

Constructors

SetProjectBranchHeadResponseNotFound

SetProjectBranchHeadResponseExpectedCausalHashMismatch !Hash32 !Hash32
(expected, actual)

SetProjectBranchHeadResponsePublishedReleaseIsImmutable

SetProjectBranchHeadResponseDeprecatedReleaseIsImmutable

SetProjectBranchHeadResponseSuccess

setProjectBranchHead :: SetProjectBranchHeadRequest -> Cli SetProjectBranchHeadResponse

Set a project branch head (can be a fast-forward or force-push).

Temporary special hard-coded base url

hardCodedBaseUrl :: BaseUrl

hardCodedUri :: URI