Unison.Cli.Share.Projects.Types

Description

Types of Unison.Cli.Share.Projects, put in their own module to avoid cyclic dependencies.

Refer to that module's documentation, and use it instead if you can. It re-exports these types.

Documentation

data RemoteProject

A remote project.

Constructors

RemoteProject

Fields

projectId :: RemoteProjectId

projectName :: ProjectName

latestRelease :: Maybe Semver

data RemoteProjectBranch

A remote project branch.

Constructors

RemoteProjectBranch

Fields

projectId :: RemoteProjectId

projectName :: ProjectName

branchId :: RemoteProjectBranchId

branchName :: ProjectBranchName

branchHead :: HashJWT

squashedBranchHead :: Maybe HashJWT