Unison.Cli.UniqueTypeGuidLookup

Description

This module contains functionality related to computing a "unique type guid lookup" function, which resolves a name to a unique type's GUID to reuse.

Documentation

loadUniqueTypeGuid :: ProjectPath -> Name -> Transaction (Maybe Text)