Unison.Cli.UpdateUtils

Description

This module contains functionality that is common to the general idea of "updating" a term in Unison, which is when we reassign a name from one hash to another and then see if all dependents still typecheck.

This occurs in the pull, merge, update, and upgrade commands.

Getting dependents in a namespace

getNamespaceDependentsOf :: Defns (BiMultimap Referent Name) (BiMultimap TypeReference Name) -> DefnsF Set TermReference TypeReference -> Transaction (DefnsF (Map Name) TermReferenceId TypeReferenceId)

Given an unconflicted namespace and a set of dependencies, return the subset of the namespace that consists of only the (transitive) dependents of the dependencies.

subtractDependents :: DefnsF Set TermReferenceId TypeReferenceId -> DefnsF (Map Name) Referent TypeReference -> DefnsF (Map Name) Referent TypeReference

Hydrating definitions

hydrateRefs :: forall (m :: Type -> Type) v a. Codebase m v a -> DefnsF Set TermReferenceId TypeReferenceId -> Transaction (Defns (Map TermReferenceId (Term v a, Type v a)) (Map TypeReferenceId (Decl v a)))

Hydrate termtype references to actual termstypes.

nameHydratedRefIds :: DefnsF (Map name) TermReferenceId TypeReferenceId -> Defns (Map TermReferenceId term) (Map TypeReferenceId typ) -> DefnsF (Map name) (TermReferenceId, term) (TypeReferenceId, typ)

Associate names with hydrated terms/types.

nameHydratedRefIds2 :: Ord name => Defns (BiMultimap Referent name) (BiMultimap TypeReference name) -> Defns (Map TermReferenceId term) (Map TypeReferenceId typ) -> DefnsF (Map name) (TermReferenceId, term) (TypeReferenceId, typ)

Like nameHydratedRefIds, but takes the entire namespace as a first argument, which includes constructors.

Unique type guids

makeUniqueTypeGuids :: Map Name TypeReference -> Transaction (Map Name Text)

Parsing and typechecking

parseAndTypecheck :: Pretty ColorText -> ParsingEnv Transaction -> Cli (Maybe (TypecheckedUnisonFile Symbol Ann))