Unison.Codebase.Editor.Command

Documentation

data Command a where

Constructors

AskEnv :: Command Env

LocalEnv :: (Env -> Env) -> Free Command a -> Command a

GetLoopState :: Command LoopState

PutLoopState :: LoopState -> Command ()

Eval :: IO a -> Command a

WithRunInIO :: ((forall x. Action x -> IO x) -> IO a) -> Command a

WithResource :: (forall r. (a -> IO r) -> IO r) -> Command a

Reset :: Action a -> Command a

Abort :: Command a

Quit :: Command a

data Env

Constructors

Env

Fields

authHTTPClient :: AuthenticatedHttpClient

codebase :: Codebase IO Symbol Ann

config :: Config

credentialManager :: CredentialManager

generateUniqueName :: IO UniqueName

Generate a unique name.

loadSource :: SourceName -> IO LoadSourceResult

How to load source code.

notify :: Output -> IO ()

What to do with output for the user.

notifyNumbered :: NumberedOutput -> IO NumberedArgs

What to do with numbered output for the user.

runtime :: Runtime Symbol

sandboxedRuntime :: Runtime Symbol

serverBaseUrl :: Maybe BaseUrl

ucmVersion :: UCMVersion

data LoopState

Constructors

LoopState

Fields

_root :: Branch IO

_lastSavedRoot :: Branch IO

_currentPathStack :: NonEmpty Absolute

_latestFile :: Maybe (FilePath, SkipNextUpdate)

_latestTypecheckedFile :: Maybe (TypecheckedUnisonFile Symbol Ann)

_lastInput :: Maybe Input

_numberedArgs :: NumberedArgs

type AmbientAbilities v = [Type v Ann]

type LexedSource = (Text, [Token Lexeme])

type Source = Text

type SourceName = Text

type TypecheckingResult v = Result (Seq (Note v Ann)) (Either Names (TypecheckedUnisonFile v Ann))

data LoadSourceResult

Constructors

InvalidSourceNameError

LoadError

LoadSuccess Text

type UseCache = Bool

type EvalResult v = ([(v, Term v ())], Map v (Ann, WatchKind, Id, Term v (), Term v (), IsCacheHit))

commandName :: Command a -> String

lookupEvalResult :: Ord v => v -> EvalResult v -> Maybe (Term v ())

abort :: Action a

quit :: Action a

with :: (forall b. (a -> IO b) -> IO b) -> Action a

Acquire a resource for the duration of an action.

reset :: Action a -> Action a

Delimit the scope of inner with actions.

root :: Lens' LoopState (Branch IO)

numberedArgs :: Lens' LoopState NumberedArgs

currentPathStack :: Lens' LoopState (NonEmpty Absolute)

lastInput :: Lens' LoopState (Maybe Input)

lastSavedRoot :: Lens' LoopState (Branch IO)

latestFile :: Lens' LoopState (Maybe (FilePath, SkipNextUpdate))

latestTypecheckedFile :: Lens' LoopState (Maybe (TypecheckedUnisonFile Symbol Ann))

currentPath :: Getter LoopState Absolute

loopState0 :: Branch IO -> Absolute -> LoopState

respond :: Output -> Action ()

respondNumbered :: NumberedOutput -> Action ()

askCodebase :: Action (Codebase IO Symbol Ann)

Get the codebase out of the environment.

askGenerateUniqueName :: Action (IO UniqueName)

Get an action that generates a unique name out of the environment.

askLoadSource :: Action (SourceName -> IO LoadSourceResult)

Get how to load source code out of the environment.

askRuntime :: Action (Runtime Symbol)

Get the runtime out of the environment.

askSandboxedRuntime :: Action (Runtime Symbol)

Get the sandboxed runtime out of the environment.

askServerBaseUrl :: Action (Maybe BaseUrl)

Get the server base url out of the environment.

getConfig :: Configured a => Text -> Action (Maybe a)

Lookup a config value by key.

type InputDescription = Text

newtype Action a

Constructors

Action

Fields

unAction :: Free Command a

eval :: Command a -> Action a