Unison.Codebase.Editor.HandleCommand

Documentation

data ReturnType a

Constructors

Success a

HaltStep

HaltRepl

data Bailing

Constructors

HaltingStep

HaltingRepl

newtype Cli r a

Constructors

Cli

Fields

unCli :: (a -> Env -> IO (ReturnType r)) -> Env -> IO (ReturnType r)

withCliToIO' :: forall r a. ((forall m x. (m x -> Cli x x) -> m x -> IO x) -> IO a) -> Cli r a

withCliToIO generalized to accept other monads that we can turn into Cli (e.g. Action)

withCliToIO :: ((forall x. Cli x x -> IO x) -> IO a) -> Cli r a

Provide a way to run Cli to IO. Note that this also delimits the scope of and succeedWith or with calls.

short :: ReturnType r -> Cli r a

succeedWith :: r -> Cli r a

Short-circuit success. Returns r to the nearest enclosing scopeWith

abortStep :: Cli r a

Short-circuit success

haltRepl :: Cli r a

Halt the repl

with :: (forall x. (a -> IO x) -> IO x) -> Cli r a

Wrap a continuation with Cli. Provides a nicer syntax to resource acquiring functions.

resource <- with (bracket acquire close)

Delimit the scope of acquired resources with scopeWith.

scopeWith :: Cli x x -> Cli r x

Delimit the scope of with calls

commandLine :: Env -> LoopState -> IO (Either Event Input) -> (Either Event Input -> Action ()) -> IO (Maybe (), LoopState)