Unison.Codebase.Editor.HandleInput.Branch

Description

branch input handler

Documentation

data CreateFrom

Constructors

CreateFrom'ParentBranch ProjectBranch

CreateFrom'Namespace (Branch IO)

CreateFrom'CausalHash CausalHash

CreateFrom'MergeParents (CreateFromMergeSource, CausalHash, Map Name Text) (ProjectBranch, CausalHash, Map Name Text) (Branch Transaction)

CreateFrom'Update (ProjectBranch, CausalHash, Map Name Text) (Branch IO)

CreateFrom'Upgrade (ProjectBranch, CausalHash, Map Name Text) (Branch IO)

CreateFrom'Nothingness

data CreateFromMergeSource

Constructors

CreateFromMergeSource'Local ProjectBranch

CreateFromMergeSource'Remote RemoteProjectBranch URI

CreateFromMergeSource'LooseCode

handleBranch :: BranchSourceI -> ProjectAndBranch (Maybe ProjectName) ProjectBranchName -> Cli ()

Create a new project branch from an existing project branch or namespace.

createBranch :: Text -> CreateFrom -> Project -> Transaction ProjectBranchName -> Cli (ProjectAndBranch ProjectId ProjectBranchId, ProjectBranchName)

createBranch description createFrom project getNewBranchName:

Creates a new branch row in project at the name from getNewBranchName (failing if branch already exists in project).

Switches to the new branch.

This bit of functionality is factored out from the main handleBranch handler because it is also called by the release.draft command, which essentially just creates a branch, but with some different output for the user.

Returns the branch id and name of the newly-created branch.