Unison.Codebase.Editor.HandleInput.DeleteBranch

Description

delete.branch input handler

Documentation

handleDeleteBranch :: ProjectAndBranch (Maybe ProjectName) ProjectBranchName -> Cli ()

Delete a project branch.

Currently, deleting a branch means deleting its project_branch row, then deleting its contents from the namespace. Its children branches, if any, are reparented to their grandparent, if any. You may delete the only branch in a project.

handleDeleteBranch2 :: ProjectAndBranch Project ProjectBranch -> Cli ()

Like handleDeleteBranch2, but for when the branch name to delete is already resolved to a branch.

doDeleteProjectBranch :: HasCallStack => ProjectAndBranch Project ProjectBranch -> Cli ()

Delete a project branch and record an entry in the reflog.