Unison.Codebase.Editor.HandleInput.EditNamespace

Documentation

handleEditNamespace :: OutputLocation -> [Path'] -> Cli ()

getNamesForEdit :: forall (m :: Type -> Type). Codebase m Symbol Ann -> PrettyPrintEnvDecl -> Names -> Transaction (Map TypeReference (DisplayObject () (Decl Symbol Ann)), Map TermReference (DisplayObject (Type Symbol Ann) (Term Symbol Ann)))

Get names "for edit": gets types and terms out the codebase as display objects, but is careful not to get an auto-generated record accessor term like `Foo.bar.set` if it's also getting the corresponding type Foo. This is because these name are "for edit", i.e. going into a scratch file, where parsing the record type will generate its accessors.