Unison.Codebase.Editor.HandleInput.MoveTo

Documentation

handleMoveTo :: NonEmpty Path' -> Path' -> Text -> Cli ()

Move one or more items INTO a destination namespace. For example: `moveTo foo.bar baz` moves `foo.bar` to `baz.bar` (keeps the final segment) `moveTo foo bar baz dest` moves all three into `dest.foo`, `dest.bar`, `dest.baz`

If multiple sources have the same final segment, we move the non-conflicting items and report the conflict to the user.