Unison.Codebase.Editor.HandleInput.NamespaceDependencies

Documentation

namespaceDependencies :: forall m i v. Ord v => Branch0 m -> Action m i v (Map LabeledDependency (Set Name))

Check the dependencies of all types, terms, and metadata in the current namespace, returns a map of dependencies which do not have a name within the current namespace, alongside the names of all of that thing's dependents.

This is non-transitive, i.e. only the first layer of external dependencies is returned.

So if my namespace depends on .base.Bag.map; which depends on base.Map.mapKeys, only .base.Bag.map is returned unless some other definition inside my namespace depends on base.Map.mapKeys directly.

Returns a Set of names rather than using the PPE since we already have the correct names in scope on this branch, and also want to list ALL names of dependents, including aliases.