Unison.Codebase.Editor.HandleInput.Pull

Description

pull input handler

Documentation

handlePull :: PullSourceTarget -> PullMode -> Cli ()

loadPropagateDiffDefaultPatch :: Text -> Maybe (Either ProjectPath (ProjectAndBranch Project ProjectBranch)) -> ProjectPath -> Cli ()

mergeBranchAndPropagateDefaultPatch :: MergeMode -> Text -> Maybe Output -> Branch IO -> Maybe (Either ProjectPath (ProjectAndBranch Project ProjectBranch)) -> ProjectPath -> Cli ()

supply dest0 if you want to print diff messages supply unchangedMessage if you want to display it if merge had no effect

propagatePatch :: Text -> Patch -> ProjectPath -> Cli Bool