Unison.Codebase.Editor.HandleInput.Update2

Documentation

handleUpdate2 :: Cli ()

Misc helpers to be organized later

addDefinitionsToUnisonFile :: (forall void. Output -> Transaction void) -> Codebase IO Symbol Ann -> (Maybe Int -> Name -> Either Output [Name]) -> (Relation Name TermReferenceId, Relation Name TypeReferenceId) -> UnisonFile Symbol Ann -> Transaction (UnisonFile Symbol Ann)

addDefinitionsToUnisonFile abort codebase doFindCtorNames definitions file adds all definitions to file, avoiding overwriting anything already in file. Every definition is put into the file with every naming it has in names "on the left-hand-side of the equals" (but yes type decls don't really have a LHS).

TODO: find a better module for this function, as it's used in a couple places

findCtorNames :: UpdateOrUpgrade -> Names -> Map ForwardName (Referent, Name) -> Maybe Int -> Name -> Either Output [Name]

given a decl name, find names for all of its constructors, in order.

Precondition: n is an element of names

findCtorNamesMaybe :: UpdateOrUpgrade -> Names -> Map ForwardName (Referent, Name) -> Maybe Int -> Name -> Either Output (Maybe [Name])

forwardCtorNames :: Names -> Map ForwardName (Referent, Name)

O(r + c * d) touches all the referents (r), and all the NameSegments (d) of all of the Con referents (c)

makeParsingEnv :: Absolute -> Names -> Cli (ParsingEnv Transaction)

prettyParseTypecheck :: UnisonFile Symbol Ann -> PrettyPrintEnvDecl -> ParsingEnv Transaction -> Cli (Either (Pretty ColorText) (TypecheckedUnisonFile Symbol Ann))

typecheckedUnisonFileToBranchUpdates

Arguments

:: (forall void. Output -> Transaction void)

-> (Name -> Either Output (Maybe [Name]))
Returns Nothing if the decl isn't in namesExcludingLibdeps, in which case we know the decl is new and do not need to generate delete actions for it.

-> TypecheckedUnisonFile Symbol Ann

-> Transaction [(Path, Branch0 m -> Branch0 m)]

getNamespaceDependentsOf :: Names -> Set Reference -> Transaction (Relation Name TermReferenceId, Relation Name TypeReferenceId)

Given a namespace and a set of dependencies, return the subset of the namespace that consists of only the (transitive) dependents of the dependencies.

makeComplicatedPPE :: Int -> Names -> Names -> (Relation Name TermReferenceId, Relation Name TypeReferenceId) -> PrettyPrintEnvDecl