Unison.Codebase.Editor.Input

Documentation

data Input

Constructors

ForkLocalBranchI (Either ShortCausalHash BranchRelativePath) BranchRelativePath

MergeLocalBranchI LooseCodeOrProject LooseCodeOrProject MergeMode

PreviewMergeLocalBranchI LooseCodeOrProject LooseCodeOrProject

DiffNamespaceI BranchId BranchId

PullRemoteBranchI PullSourceTarget SyncMode PullMode Verbosity

PushRemoteBranchI PushRemoteBranchInput

ResetRootI (Either ShortCausalHash Path')

ResetI (These (Either ShortCausalHash Path') (ProjectAndBranch (Maybe ProjectName) ProjectBranchName)) (Maybe LooseCodeOrProject)

CreateMessage (Pretty ColorText)

SwitchBranchI Path'

UpI

PopBranchI

NamesI IsGlobal (HashQualified Name)

AliasTermI HashOrHQSplit' Split'

AliasTypeI HashOrHQSplit' Split'

AliasManyI [HQSplit] Path'

MoveAllI Path' Path'

MoveTermI HQSplit' Split'

MoveTypeI HQSplit' Split'

MoveBranchI Path' Path'

MovePatchI Split' Split'

CopyPatchI Split' Split'

DeleteI DeleteTarget

LoadI (Maybe FilePath)

ClearI

AddI (Set Name)

PreviewAddI (Set Name)

UpdateI OptionalPatch (Set Name)

Update2I

PreviewUpdateI (Set Name)

TodoI (Maybe PatchPath) Path'

PropagatePatchI PatchPath Path'

ListEditsI (Maybe PatchPath)

DeprecateTermI PatchPath HQSplit'

DeprecateTypeI PatchPath HQSplit'

ReplaceI (HashQualified Name) (HashQualified Name) (Maybe PatchPath)

RemoveTermReplacementI (HashQualified Name) (Maybe PatchPath)

RemoveTypeReplacementI (HashQualified Name) (Maybe PatchPath)

UndoI

HistoryI (Maybe Int) (Maybe Int) BranchId

ExecuteI Text [String]

SaveExecuteResultI Name

IOTestI (HashQualified Name)

IOTestAllI

MakeStandaloneI String (HashQualified Name)

ExecuteSchemeI Text [String]

CompileSchemeI Text (HashQualified Name)

TestI TestInput

CreateAuthorI NameSegment Text

DisplayI OutputLocation (NonEmpty (HashQualified Name))

DocsI (NonEmpty HQSplit')

FindI Bool FindScope [String]

FindShallowI Path'

FindPatchI

StructuredFindI FindScope (HashQualified Name)

StructuredFindReplaceI (HashQualified Name)

ShowDefinitionI OutputLocation ShowDefinitionScope (NonEmpty (HashQualified Name))

ShowDefinitionByPrefixI OutputLocation [HashQualified Name]

ShowReflogI

UpdateBuiltinsI

MergeBuiltinsI

MergeIOBuiltinsI

ListDependenciesI (HashQualified Name)

ListDependentsI (HashQualified Name)

NamespaceDependenciesI (Maybe Path')
List all external dependencies of a given namespace, or the current namespace if no path is provided.

DebugTabCompletionI [String]

DebugFuzzyOptionsI String [String]

DebugFormatI

DebugNumberedArgsI

DebugTypecheckedUnisonFileI

DebugDumpNamespacesI

DebugDumpNamespaceSimpleI

DebugTermI Bool (HashQualified Name)

DebugTypeI (HashQualified Name)

DebugLSPFoldRangesI

DebugClearWatchI

DebugDoctorI

DebugNameDiffI ShortCausalHash ShortCausalHash

QuitI

ApiI

UiI Path'

DocToMarkdownI Name

DocsToHtmlI Path' FilePath

GistI GistInput

AuthLoginI

VersionI

DiffNamespaceToPatchI DiffNamespaceToPatchInput

ProjectCreateI Bool (Maybe ProjectName)

ProjectRenameI ProjectName

ProjectSwitchI ProjectAndBranchNames

ProjectsI

BranchI BranchSourceI (ProjectAndBranch (Maybe ProjectName) ProjectBranchName)

BranchRenameI ProjectBranchName

BranchesI (Maybe ProjectName)

CloneI ProjectAndBranchNames (Maybe ProjectAndBranchNames)

ReleaseDraftI Semver

UpgradeI !NameSegment !NameSegment

EditNamespaceI [Path]

data BranchSourceI

The source of a branch command: what to make the new branch from.

Constructors

BranchSourceI'CurrentContext
Create a branch from the current context

BranchSourceI'Empty
Create an empty branch

BranchSourceI'LooseCodeOrProject LooseCodeOrProject
Create a branch from this loose-code-or-project

data DiffNamespaceToPatchInput

Constructors

DiffNamespaceToPatchInput

Fields

branchId1 :: BranchId

branchId2 :: BranchId

patch :: Split'

data GistInput

"push.gist repo" pushes the contents of the current namespace to repo.

Constructors

GistInput

Fields

repo :: WriteGitRepo

data PullSourceTarget

Pull source and target: either neither is specified, or only a source, or both.

Constructors

PullSourceTarget0

PullSourceTarget1 (ReadRemoteNamespace (These ProjectName ProjectBranchNameOrLatestRelease))

PullSourceTarget2 (ReadRemoteNamespace (These ProjectName ProjectBranchNameOrLatestRelease)) LooseCodeOrProject

data PushRemoteBranchInput

Constructors

PushRemoteBranchInput

Fields

sourceTarget :: PushSourceTarget

pushBehavior :: PushBehavior

syncMode :: SyncMode

data PushSourceTarget

Push source and target: either neither is specified, or only a target, or both.

Constructors

PushSourceTarget0

PushSourceTarget1 (WriteRemoteNamespace (These ProjectName ProjectBranchName))

PushSourceTarget2 PushSource (WriteRemoteNamespace (These ProjectName ProjectBranchName))

data PushSource

Constructors

PathySource Path'

ProjySource (These ProjectName ProjectBranchName)

data TestInput

Constructors

TestInput

Fields

includeLibNamespace :: Bool

Should we run tests in the lib namespace?

showFailures :: Bool

showSuccesses :: Bool

data Event

Constructors

UnisonFileChanged SourceName Source

IncomingRootBranch (Set CausalHash)

data OutputLocation

Constructors

ConsoleLocation

LatestFileLocation

FileLocation FilePath

type PatchPath = Split'

type BranchId = Either ShortCausalHash Path'

type AbsBranchId = Either ShortCausalHash Absolute

type LooseCodeOrProject = These Path' (ProjectAndBranch (Maybe ProjectName) ProjectBranchName)

A lot of commands can take either a loose code path or a project branch in the same argument slot. Usually, those have distinct syntaxes, but sometimes it's ambiguous, in which case we'd parse a These. The command itself can decide what to do with the ambiguity.

parseBranchId :: String -> Either Text BranchId

parseBranchId2 :: String -> Either (Pretty ColorText) (Either ShortCausalHash BranchRelativePath)

parseShortCausalHash :: String -> Either String ShortCausalHash

type HashOrHQSplit' = Either ShortHash HQSplit'

data Insistence

Should we force the operation or not?

Constructors

Force

Try

data PullMode

Constructors

PullWithHistory

PullWithoutHistory

data OptionalPatch

Constructors

NoPatch

DefaultPatch

UsePatch PatchPath

data FindScope

Constructors

FindLocal

FindLocalAndDeps

FindGlobal

data ShowDefinitionScope

Constructors

ShowDefinitionLocal

ShowDefinitionGlobal

type IsGlobal = Bool

data DeleteOutput

Constructors

DeleteOutput'Diff

DeleteOutput'NoDiff

data DeleteTarget

Constructors

DeleteTarget'TermOrType DeleteOutput [HQSplit']

DeleteTarget'Term DeleteOutput [HQSplit']

DeleteTarget'Type DeleteOutput [HQSplit']

DeleteTarget'Namespace Insistence (Maybe Split')

DeleteTarget'Patch Split'

DeleteTarget'ProjectBranch (ProjectAndBranch (Maybe ProjectName) ProjectBranchName)

DeleteTarget'Project ProjectName