Unison.Codebase.Editor.Slurp

Documentation

data SlurpOp

The operation which is being performed or checked.

Constructors

AddOp

UpdateOp

CheckOp

slurpFile :: TypecheckedUnisonFile Symbol Ann -> Set Symbol -> SlurpOp -> Names -> SlurpResult

Analyze a file and determine the status of all of its definitions with respect to a set of vars to analyze and an operation you wish to perform.