Unison.Codebase.Transcript

Description

The data model for Unison transcripts.

Documentation

type ExpectingError = Bool

type HasBug = Bool

type ScratchFileName = Text

data Hidden

Constructors

Shown

HideOutput

HideAll

data UcmLine

Constructors

UcmCommand UcmContext Text

UcmComment Text
Text does not include the '--' prefix.

UcmOutputLine Text

data UcmContext

Where a command is run: a project branch (myproject/mybranch>).

Constructors

UcmContextEmpty
Use the current project & branch

UcmContextProject (ProjectAndBranch ProjectName ProjectBranchName)
Explicit project & branch

data APIRequest

Constructors

GetRequest Text

PostRequest Text Text
URL, Body

APIComment Text

APIResponse Text

pattern CMarkCodeBlock :: Maybe PosInfo -> Text -> Text -> Node

type Stanza = Either Node ProcessedBlock

data InfoTags a

Constructors

InfoTags

Fields

hidden :: Maybe Hidden

Nothing represents the default value, which differs depending on the TranscriptType and ProcessedBlock tag.

expectingError :: ExpectingError

hasBug :: HasBug

generated :: Bool

additionalTags :: a

getCommonInfoTags :: ProcessedBlock -> InfoTags ()

data Transcript

A transcript is executable Markdown, akin to a Jupyter Notebook.

NB: This is effectively the AST. It needs to preserve anything we want to be able to serialize into transcript output.

Constructors

Transcript

Fields

frontmatter :: Value

stanzas :: [Stanza]

settings :: Transcript -> Settings

data TranscriptType

Different types make various settingal changes.

Constructors

Standard
The default type, nothing changed from the default setting.

Tutorial
Enables auto-update and changes the default hidden value for Unison blocks to HideOutput.

data Behaviors (f :: Type -> Type)

The atomic behaviors that can be controlled by settings. They may be set individually, or implied by some aggregate setting.

Constructors

Behaviors

Fields

autoupdate :: f Bool

getHidden :: f (ProcessedBlock -> Hidden)

extractBehaviors :: Settings -> Behaviors Identity

data Settings

A set of customizable settings for a transcript.

Constructors

Settings

Fields

transcriptType :: Maybe TranscriptType

behaviors :: Behaviors Maybe

defaultInfoTags :: a -> InfoTags a

data ProcessedBlock

Constructors

Ucm (InfoTags ()) [UcmLine]

Unison (InfoTags (Maybe ScratchFileName)) Text

API (InfoTags ()) [APIRequest]

data Node