Unison.Codebase.Transcript.Runner

Description

Execute transcripts.

Documentation

data Error

Constructors

ParseError (ParseErrorBundle Text Void)

RunFailure Transcript

PortBindingFailure

Exception SomeException

data Config

Constructors

Config

Fields

terminalWidth :: Width

fzfPath :: Maybe FilePath

Nothing uses the default, a value overrides $FZF_PATH. It may be set to `NONE` to disable the use of FZF.

credentialsFile :: Maybe FilePath

isTest :: Bool

Control additional test-related values.

TODO: This should be broken down with the individual pieces included here.

defaultConfig :: Config

A reasonable set of defaults to use.

render transcript errors at a width of 65 chars

use $FZF_PATH

use the default credentials path

testConfig :: FilePath -> Config

type Runner

Arguments

= String
The name of the transcript to run.

-> ByteString
The contents of the transcript to run.

-> Codebase IO Symbol Ann

-> IO (Either Error Transcript)

withRunner

Arguments

:: MonadUnliftIO m

=> Config
Whether to treat this transcript run as a transcript test, which will try to make output deterministic

-> Verbosity

-> UCMVersion

-> (Runner -> m r)

-> m r