Unison.CommandLine.Completion

Completers

exactComplete :: String -> [String] -> [Completion]

Constructs a list of Completions from a query and completion options by filtering them for prefix matches. A completion will be selected if it's an exact match for a provided option.

prefixCompleteTermOrType :: String -> ProjectPath -> Transaction [Completion]

Completes a term or type argument by prefix-matching against the query.

prefixCompleteTerm :: String -> ProjectPath -> Transaction [Completion]

Completes a term argument by prefix-matching against the query.

prefixCompleteType :: String -> ProjectPath -> Transaction [Completion]

Completes a term or type argument by prefix-matching against the query.

prefixCompletePatch :: String -> ProjectPath -> Transaction [Completion]

Completes a patch argument by prefix-matching against the query.

noCompletions :: MonadIO m => String -> Codebase m v a -> AuthenticatedHttpClient -> ProjectPath -> m [Completion]

The empty completor.

prefixCompleteNamespace :: String -> ProjectPath -> Transaction [Completion]

Completes a namespace argument by prefix-matching against the query.

prettyCompletion :: Bool -> (String, Pretty ColorText) -> Completion

fixupCompletion :: String -> [Completion] -> [Completion]

haskelineTabComplete :: MonadIO m => Map String InputPattern -> Codebase m v a -> AuthenticatedHttpClient -> ProjectPath -> CompletionFunc m

A completion func for use with Haskeline

sharePathCompletion :: MonadIO m => AuthenticatedHttpClient -> String -> m [Completion]