Unison.CommandLine.FZFResolvers

Documentation

data FZFResolver

Constructors

FZFResolver

Fields

getOptions :: OptionFetcher

definitionOptions :: OptionFetcher

Select a definition from the given branch.

termDefinitionOptions :: OptionFetcher

Select a term definition from the given branch. Returned names will match the provided Position type.

typeDefinitionOptions :: OptionFetcher

Select a type definition from the given branch. Returned names will match the provided Position type.

namespaceOptions :: OptionFetcher

Select a namespace from the given branch. Returned Path's will match the provided Position type.

projectDependencyResolver :: FZFResolver

projectNameOptions :: OptionFetcher

All possible local project names E.g. '@unison/base'

projectBranchOptions :: OptionFetcher

All possible local project/branch names. E.g. '@unisonbasemain'

projectBranchOptionsWithinCurrentProject :: OptionFetcher

All possible local branch names within the current project. E.g. '@unisonbasemain'

fuzzySelectFromList :: [Text] -> FZFResolver

Select a namespace from the given branch. Returned Path's will match the provided Position type.

multiResolver :: [OptionFetcher] -> FZFResolver

Combine multiple option fetchers into one resolver.

definitionResolver :: FZFResolver

typeDefinitionResolver :: FZFResolver

termDefinitionResolver :: FZFResolver

namespaceResolver :: FZFResolver

namespaceOrDefinitionResolver :: FZFResolver

projectAndOrBranchArg :: FZFResolver

A project name, branch name, or both.

projectOrBranchResolver :: FZFResolver

projectBranchResolver :: FZFResolver

projectBranchWithinCurrentProjectResolver :: FZFResolver

projectNameResolver :: FZFResolver

fuzzySelectHeader :: Text -> Text

Exported from here just so the debug command and actual implementation can use the same messaging.

>>> fuzzySelectHeader "definition to view"
"Select a definition to view:"

>>> fuzzySelectHeader "alias name"
"Select an alias name:"