Unison.CommandLine.FuzzySelect

Description

Command-line fuzzy selection of arbitrary values. Shells out to fzf for the actual selection.

Documentation

fuzzySelect :: Options -> FuzzySelections a -> IO (Maybe [a])

Allows prompting the user to interactively fuzzy-select a result from a list of options, currently shells out to fzf under the hood. If fzf is missing, or an error (other than ctrl-c) occurred, returns Nothing.

isFZFInstalled :: Bool

fzfPathEnvVar :: String

An environment variable that can be set to override the default fzf executable.

data Options

Fuzzy Selection options

Constructors

Options

Fields

allowMultiSelect :: Bool

data FuzzySelections a where

Constructors

SelectFromChoices :: forall a. (a -> Text) -> [a] -> FuzzySelections a

SelectFiles :: FuzzySelections Text

defaultOptions :: Options

Default Options