Unison.CommandLine.InputPatterns

Documentation

showPatternHelp :: InputPattern -> Pretty ColorText

patternName :: InputPattern -> Pretty ColorText

makeExample :: InputPattern -> [Pretty ColorText] -> Pretty ColorText

makeExampleNoBackticks :: InputPattern -> [Pretty ColorText] -> Pretty ColorText

makeExample' :: InputPattern -> Pretty ColorText

makeExampleEOS :: InputPattern -> [Pretty ColorText] -> Pretty ColorText

helpFor :: InputPattern -> Pretty ColorText

mergeBuiltins :: InputPattern

mergeIOBuiltins :: InputPattern

updateBuiltins :: InputPattern

todo :: InputPattern

load :: InputPattern

clear :: InputPattern

add :: InputPattern

previewAdd :: InputPattern

update :: InputPattern

updateOldNoPatch :: InputPattern

updateOld :: InputPattern

previewUpdate :: InputPattern

patch :: InputPattern

view :: InputPattern

viewGlobal :: InputPattern

display :: InputPattern

displayTo :: InputPattern

docs :: InputPattern

api :: InputPattern

ui :: InputPattern

undo :: InputPattern

viewByPrefix :: InputPattern

sfind :: InputPattern

sfindReplace :: InputPattern

find :: InputPattern

findAll :: InputPattern

findGlobal :: InputPattern

find' :: String -> FindScope -> InputPattern

findShallow :: InputPattern

findVerbose :: InputPattern

findVerboseAll :: InputPattern

findPatch :: InputPattern

renameTerm :: InputPattern

moveAll :: InputPattern

renameType :: InputPattern

deleteGen :: Maybe String -> ArgumentType -> String -> ([HQSplit'] -> DeleteTarget) -> InputPattern

delete :: InputPattern

deleteVerbose :: InputPattern

deleteTerm :: InputPattern

deleteTermVerbose :: InputPattern

deleteType :: InputPattern

deleteTypeVerbose :: InputPattern

deleteTermReplacementCommand :: String

deleteTypeReplacementCommand :: String

deleteReplacement :: Bool -> InputPattern

deleteProject :: InputPattern

deleteBranch :: InputPattern

deleteTermReplacement :: InputPattern

deleteTypeReplacement :: InputPattern

aliasTerm :: InputPattern

aliasType :: InputPattern

aliasMany :: InputPattern

up :: InputPattern

cd :: InputPattern

back :: InputPattern

deleteNamespace :: InputPattern

deleteNamespaceForce :: InputPattern

deleteNamespaceParser :: Pretty ColorText -> Insistence -> [String] -> Either (Pretty ColorText) Input

deletePatch :: InputPattern

movePatch :: String -> String -> Either (Pretty ColorText) Input

copyPatch' :: String -> String -> Either (Pretty ColorText) Input

copyPatch :: InputPattern

renamePatch :: InputPattern

renameBranch :: InputPattern

history :: InputPattern

forkLocal :: InputPattern

reset :: InputPattern

resetRoot :: InputPattern

pull :: InputPattern

pullVerbose :: InputPattern

pullWithoutHistory :: InputPattern

pullImpl :: String -> [String] -> Verbosity -> PullMode -> Pretty ColorText -> InputPattern

pullExhaustive :: InputPattern

debugTabCompletion :: InputPattern

debugFuzzyOptions :: InputPattern

debugFormat :: InputPattern

push :: InputPattern

pushCreate :: InputPattern

pushForce :: InputPattern

pushExhaustive :: InputPattern

squashMerge :: InputPattern

mergeLocal :: InputPattern

parseLooseCodeOrProject :: String -> Maybe LooseCodeOrProject

diffNamespace :: InputPattern

previewMergeLocal :: InputPattern

replaceEdit :: (HashQualified Name -> HashQualified Name -> Maybe PatchPath -> Input) -> InputPattern

replace :: InputPattern

viewReflog :: InputPattern

edit :: InputPattern

editNamespace :: InputPattern

topicNameArg :: ArgumentType

codebaseServerNameArg :: ArgumentType

helpTopics :: InputPattern

helpTopicsMap :: Map String (Pretty ColorText)

help :: InputPattern

quit :: InputPattern

viewPatch :: InputPattern

names :: IsGlobal -> InputPattern

dependents :: InputPattern

dependencies :: InputPattern

namespaceDependencies :: InputPattern

debugNumberedArgs :: InputPattern

debugFileHashes :: InputPattern

debugDumpNamespace :: InputPattern

debugDumpNamespaceSimple :: InputPattern

debugTerm :: InputPattern

debugTermVerbose :: InputPattern

debugType :: InputPattern

debugLSPFoldRanges :: InputPattern

debugClearWatchCache :: InputPattern

debugDoctor :: InputPattern

debugNameDiff :: InputPattern

test :: InputPattern

testAll :: InputPattern

docsToHtml :: InputPattern

docToMarkdown :: InputPattern

execute :: InputPattern

saveExecuteResult :: InputPattern

ioTest :: InputPattern

ioTestAll :: InputPattern

makeStandalone :: InputPattern

runScheme :: InputPattern

compileScheme :: InputPattern

createAuthor :: InputPattern

gist :: InputPattern

authLogin :: InputPattern

printVersion :: InputPattern

diffNamespaceToPatch :: InputPattern

projectCreate :: InputPattern

projectCreateEmptyInputPattern :: InputPattern

projectRenameInputPattern :: InputPattern

projectSwitch :: InputPattern

projectsInputPattern :: InputPattern

branchesInputPattern :: InputPattern

branchInputPattern :: InputPattern

branchEmptyInputPattern :: InputPattern

branchRenameInputPattern :: InputPattern

clone :: InputPattern

releaseDraft :: InputPattern

upgrade :: InputPattern

validInputs :: [InputPattern]

patternMap :: Map String InputPattern

A map of all command patterns by pattern name or alias.

visibleInputs :: [InputPattern]

commandNames :: [String]

commandNameArg :: ArgumentType

exactDefinitionArg :: ArgumentType

fuzzyDefinitionQueryArg :: ArgumentType

definitionQueryArg :: ArgumentType

exactDefinitionTypeQueryArg :: ArgumentType

exactDefinitionTypeOrTermQueryArg :: ArgumentType

exactDefinitionTermQueryArg :: ArgumentType

patchArg :: ArgumentType

namespaceArg :: ArgumentType

namespaceOrProjectBranchArg :: ProjectBranchSuggestionsConfig -> ArgumentType

Usually you'll want one or the other, but some commands like merge support both right now.

namespaceOrDefinitionArg :: ArgumentType

dependencyArg :: ArgumentType

A dependency name. E.g. if your project has `lib.base`, base would be a dependency name.

newNameArg :: ArgumentType

noCompletionsArg :: ArgumentType

filePathArg :: ArgumentType

gitUrlArg :: ArgumentType

remoteNamespaceArg :: ArgumentType

Refers to a namespace on some remote code host.

data ProjectInclusion

Constructors

OnlyWithinCurrentProject

OnlyOutsideCurrentProject

AllProjects

data BranchInclusion

Constructors

ExcludeCurrentBranch

AllBranches

projectsByPrefix :: MonadIO m => ProjectInclusion -> Codebase m v a -> Absolute -> Text -> m [(ProjectId, ProjectName)]

data ProjectBranchSuggestionsConfig

Constructors

ProjectBranchSuggestionsConfig

Fields

showProjectCompletions :: Bool

projectInclusion :: ProjectInclusion

branchInclusion :: BranchInclusion

projectAndOrBranchSuggestions :: MonadIO m => ProjectBranchSuggestionsConfig -> String -> Codebase m v a -> AuthenticatedHttpClient -> Absolute -> m [Completion]

projectToCompletion :: Project -> Completion

projectBranchToCompletion :: ProjectName -> (ProjectBranchId, ProjectBranchName) -> Completion

handleBranchesComplete :: MonadIO m => ProjectBranchSuggestionsConfig -> Text -> Codebase m v a -> Absolute -> m [Completion]

filterBranches :: ProjectBranchSuggestionsConfig -> Absolute -> [(ProjectBranchId, a)] -> [(ProjectBranchId, a)]

currentProjectBranchToCompletion :: (ProjectBranchId, ProjectBranchName) -> Completion

branchRelativePathSuggestions :: MonadIO m => ProjectBranchSuggestionsConfig -> String -> Codebase m v a -> AuthenticatedHttpClient -> Absolute -> m [Completion]

projectAndBranchNamesArg :: ProjectBranchSuggestionsConfig -> ArgumentType

A project name, branch name, or both.

projectBranchNameArg :: ProjectBranchSuggestionsConfig -> ArgumentType

A project branch name.

projectBranchNameWithOptionalProjectNameArg :: ArgumentType

branchRelativePathArg :: ArgumentType

projectNameArg :: ArgumentType

A project name.

data OptionalSlash

Constructors

WithSlash

NoSlash

projectNameSuggestions :: MonadIO m => OptionalSlash -> String -> Codebase m v a -> m [Completion]

parsePullSource :: Text -> Maybe (ReadRemoteNamespace (These ProjectName ProjectBranchNameOrLatestRelease))

parsePushSource :: String -> Either (Pretty ColorText) PushSource

Parse a PushSource.

parsePushTarget :: String -> Either (Pretty ColorText) (WriteRemoteNamespace (These ProjectName ProjectBranchName))

Parse a push target.

parseHashQualifiedName :: String -> Either (Pretty ColorText) (HashQualified Name)

parseWriteGitRepo :: String -> String -> Either (Pretty ColorText) WriteGitRepo

collectNothings :: (a -> Maybe b) -> [a] -> [a]

explainRemote :: PushPull -> Pretty ColorText

showErrorFancy :: ShowErrorComponent e => ErrorFancy e -> String

showErrorItem :: ErrorItem (Token Text) -> String