Unison.CommandLine.OutputMessages

Documentation

type Pretty = Pretty ColorText

shortenDirectory :: FilePath -> IO FilePath

renderFileName :: FilePath -> IO Pretty

notifyNumbered :: NumberedOutput -> (Pretty, NumberedArgs)

undoTip :: Pretty ColorText

showListEdits :: Patch -> PrettyPrintEnv -> (Pretty ColorText, NumberedArgs)

prettyURI :: URI -> Pretty

prettyReadRemoteNamespace :: ReadRemoteNamespace RemoteProjectBranch -> Pretty

prettyReadRemoteNamespaceWith :: (a -> Text) -> ReadRemoteNamespace a -> Pretty

prettyWriteRemoteNamespace :: WriteRemoteNamespace (ProjectAndBranch ProjectName ProjectBranchName) -> Pretty

notifyUser :: FilePath -> Output -> IO Pretty

expectedEmptyPushDest :: WriteRemoteNamespace Void -> Pretty

expectedNonEmptyPushDest :: WriteRemoteNamespace Void -> Pretty

prettyShareError :: ShareError -> Pretty

prettyCheckAndSetPushError :: CheckAndSetPushError -> Pretty

prettyDownloadEntitiesError :: DownloadEntitiesError -> Pretty

prettyFastForwardPathError :: Path -> FastForwardPathError -> Pretty

prettyFastForwardPushError :: FastForwardPushError -> Pretty

prettyGetCausalHashByPathError :: GetCausalHashByPathError -> Pretty

prettyPullError :: PullError -> Pretty

prettyUpdatePathError :: RepoInfo -> UpdatePathError -> Pretty

prettyUploadEntitiesError :: UploadEntitiesError -> Pretty

prettyTransportError :: CodeserverTransportError -> Pretty

invalidRepoInfo :: Text -> RepoInfo -> Pretty

pushPublicNote :: InputPattern -> Text -> [Text] -> Pretty

needDependencies :: NeedDependencies Hash32 -> Pretty

noReadPermissionForPath :: Path -> Pretty

noReadPermissionForRepo :: RepoInfo -> Pretty

noWritePermissionForPath :: Path -> Pretty

noWritePermissionForRepo :: RepoInfo -> Pretty

notFastForward :: Path -> Pretty

shareProjectNotFound :: Text -> Pretty

shareUserNotFound :: RepoInfo -> Pretty

sharePathToWriteRemotePathShare :: Path -> WriteRemoteNamespace void

shareOrigin :: Text

prettyRepoInfo :: RepoInfo -> Pretty

prettyShareLink :: WriteShareRemoteNamespace -> Pretty

prettySharePath :: Path -> Pretty

prettyFilePath :: FilePath -> Pretty

prettyPath' :: Path' -> Pretty

prettyPullTarget :: PullTarget (ProjectAndBranch Project ProjectBranch) -> Pretty

prettyBranchId :: AbsBranchId -> Pretty

prettyRelative :: Relative -> Pretty

prettyAbsolute :: Absolute -> Pretty

prettySCH :: IsString s => ShortCausalHash -> Pretty s

prettyCausalHash :: IsString s => CausalHash -> Pretty s

prettyBase32Hex :: IsString s => Base32Hex -> Pretty s

prettyBase32Hex# :: IsString s => Base32Hex -> Pretty s

prettyHash :: IsString s => Hash -> Pretty s

prettyHash32 :: IsString s => Hash32 -> Pretty s

prettyProjectName :: ProjectName -> Pretty

prettyProjectNameSlash :: ProjectName -> Pretty

prettyProjectName with a trailing slash.

prettyProjectBranchName :: ProjectBranchName -> Pretty

prettySemver :: Semver -> Pretty

prettySlashProjectBranchName :: ProjectBranchName -> Pretty

Like prettyProjectBranchName, but with a leading forward slash. This is used in some outputs to encourage/advertise an unambiguous syntax for project branches, as there's an ambiguity with single-segment relative paths.

Not all project branches are printed such: for example, when listing all branches of a project, we probably don't need or want to prefix every one with a forward slash.

prettyProjectAndBranchName :: ProjectAndBranch ProjectName ProjectBranchName -> Pretty

prettyPathOrProjectAndBranchName :: Either Path' (ProjectAndBranch ProjectName ProjectBranchName) -> Pretty

formatMissingStuff :: (Show tm, Show typ) => [(HashQualified Name, tm)] -> [(HashQualified Name, typ)] -> Pretty

displayDefinitions' :: Var v => Ord a1 => PrettyPrintEnvDecl -> Map Reference (DisplayObject () (Decl v a1)) -> Map Reference (DisplayObject (Type v a1) (Term v a1)) -> Pretty

displayRendered :: Maybe FilePath -> Pretty -> IO Pretty

displayDefinitions :: DisplayDefinitionsOutput -> IO Pretty

displayTestResults :: Bool -> PrettyPrintEnv -> [(Reference, Text)] -> [(Reference, Text)] -> Pretty

unsafePrettyTermResultSig' :: Var v => PrettyPrintEnv -> TermResult' v a -> Pretty

unsafePrettyTermResultSigFull' :: Var v => PrettyPrintEnv -> TermResult' v a -> Pretty

prettyTypeResultHeader' :: Var v => TypeResult' v a -> Pretty

prettyTypeResultHeaderFull' :: Var v => TypeResult' v a -> Pretty

prettyDeclTriple :: Var v => (HashQualified Name, Reference, DisplayObject () (Decl v a)) -> Pretty

prettyDeclPair :: Var v => PrettyPrintEnv -> (Reference, DisplayObject () (Decl v a)) -> Pretty

renderNameConflicts :: PrettyPrintEnv -> Names -> Numbered Pretty

renderEditConflicts :: PrettyPrintEnv -> Patch -> Numbered Pretty

type Numbered = State (Int, Seq String)

addNumberedArg :: String -> Numbered Int

formatNum :: Int -> Pretty

runNumbered :: Numbered a -> (a, NumberedArgs)

todoOutput :: Var v => PrettyPrintEnvDecl -> TodoOutput v a -> (Pretty, NumberedArgs)

listOfDefinitions :: Var v => FindScope -> PrettyPrintEnv -> ListDetailed -> [SearchResult' v a] -> IO Pretty

listOfLinks :: Var v => PrettyPrintEnv -> [(HashQualified Name, Maybe (Type v a))] -> IO Pretty

data ShowNumbers

Constructors

ShowNumbers

HideNumbers

showDiffNamespace :: forall v. Var v => ShowNumbers -> PrettyPrintEnv -> AbsBranchId -> AbsBranchId -> BranchDiffOutput v Ann -> (Pretty, NumberedArgs)

ppe is just for rendering type signatures `oldPath, newPath :: Path.Absolute` are just for producing fully-qualified numbered args

noResults :: FindScope -> Pretty

listOfDefinitions' :: Var v => FindScope -> PrettyPrintEnv -> ListDetailed -> [SearchResult' v a] -> Pretty

watchPrinter :: Var v => Text -> PrettyPrintEnv -> Ann -> WatchKind -> Term v () -> IsCacheHit -> Pretty

filestatusTip :: Pretty

prettyDiff :: Diff -> Pretty

prettyTermName :: PrettyPrintEnv -> Referent -> Pretty

prettyTypeName :: PrettyPrintEnv -> Reference -> Pretty

prettyReadGitRepo :: ReadGitRepo -> Pretty

prettyWriteGitRepo :: WriteGitRepo -> Pretty

prettyWhichBranchEmpty :: WhichBranchEmpty -> Pretty

Pretty-print a WhichBranchEmpty.

isTestOk :: Term v Ann -> Bool

numberedArgsForEndangerments :: PrettyPrintEnvDecl -> Map LabeledDependency (NESet LabeledDependency) -> NumberedArgs

Get the list of numbered args corresponding to an endangerment map, which is used by a few outputs. See endangeredDependentsTable.

endangeredDependentsTable :: PrettyPrintEnvDecl -> Map LabeledDependency (NESet LabeledDependency) -> Pretty ColorText

Format and render all dependents which are endangered by references going extinct.

displayBranchHash :: CausalHash -> String

Displays a full, non-truncated Branch.CausalHash to a string, e.g. #abcdef

prettyHumanReadableTime :: UTCTime -> UTCTime -> Pretty

prettyRemoteBranchInfo :: (URI, ProjectName, ProjectBranchName) -> Pretty

stripProjectBranchInfo :: Absolute -> Maybe Path

prettyAbsoluteStripProject :: Absolute -> Pretty

prettyLabeledDependencies :: PrettyPrintEnv -> Set LabeledDependency -> Pretty

listDependentsOrDependencies :: PrettyPrintEnv -> Text -> Text -> Set LabeledDependency -> [HashQualified Name] -> [HashQualified Name] -> Pretty ColorText