Unison.LSP.Conversions

Documentation

rangeToInterval :: Range -> Interval Position

annToInterval :: Ann -> Maybe (Interval Position)

uToLspPos :: Pos -> Position

Convert a Unison file-position where the first char is 1 and line is 1, to an LSP Position where the first char is 0 and line is 0.

lspToUPos :: Position -> Pos

Convert an LSP Position where the first char is 0 and line is 0, to a Unison file-position where the first char is 1 and line is 1.

uToLspRange :: Range -> Range

Convert a Unison Range where the first char is 1 and line is 1, to an LSP Range where the first char is 0 and line is 0.

lspToURange :: Range -> Range

Convert an LSP Range where the first char is 0 and line is 0, to a Unison Range where the first char is 1 and line is 1.

annToRange :: Ann -> Maybe Range

annToURange :: Ann -> Maybe Range