Unison.LSP.Diagnostics

Documentation

annToRange :: Ann -> Maybe Range

uToLspPos :: Pos -> Position

uToLspRange :: Range -> Range

infoDiagnostics :: FileAnalysis -> Lsp [Diagnostic]

noteDiagnostics :: Foldable f => Uri -> String -> f (Note Symbol Ann) -> Lsp [Diagnostic]

noteRanges :: Note Symbol Ann -> [(Range, [(Text, Range)])]

Returns a list of ranges where this note should be marked in the document, as well as a list of related ranges the note might refer to, and their relevance.

E.g. a name conflict note might mark each conflicted name, and contain references to the other conflicted name locations.

reportDiagnostics

Arguments

:: Uri

-> Maybe FileVersion

-> [Diagnostic]
Note, it's important to still send an empty list of diagnostics if there aren't any because it clears existing diagnostics in the editor

-> Lsp ()

data UnisonDiagnostic

Constructors

UnisonDiagnostic Range UnisonDiagnosticInfo

data UnisonDiagnosticInfo

Constructors

TypeError Text

NameResolutionFailure (ResolutionFailure Symbol Range)

mkDiagnostic :: Uri -> Range -> DiagnosticSeverity -> Text -> [(Text, Range)] -> Diagnostic

getSeverity :: UnisonDiagnosticInfo -> DiagnosticSeverity

diagnosticMessage :: UnisonDiagnosticInfo -> Text

toLSPDiagnostic :: UnisonDiagnostic -> Diagnostic