Unison.LSP.FileAnalysis

Documentation

checkFileByUri :: (HasUri d Uri, Lspish m) => d -> m (Maybe FileAnalysis)

Lex, parse, and typecheck a file using a VFS URI

checkFileContents :: forall (m :: Type -> Type). Lspish m => Uri -> Text -> FileVersion -> Text -> MaybeT m FileAnalysis

Lex, parse, and typecheck a file. This is split off for easier testing without needing to mock the VFS.

getFileAnalysis :: forall (m :: Type -> Type). Lspish m => Uri -> MaybeT m FileAnalysis

ppedForFile :: Lspish m => Uri -> m PrettyPrintEnvDecl

getFileSummary :: forall (m :: Type -> Type). Lspish m => Uri -> MaybeT m FileSummary

ppedForFileHelper :: Lspish m => Maybe (UnisonFile Symbol a) -> Maybe (TypecheckedUnisonFile Symbol a) -> m PrettyPrintEnvDecl

fileAnalysisWorker :: Lsp ()

getFileDefLocations :: Uri -> MaybeT Lsp (Map Symbol (Set Ann))

Get the location of user defined definitions within the file

getFileNames :: Uri -> MaybeT Lsp Names

Build a Names from a file if it's parseable.

If the file typechecks, generate names from that, otherwise, generate names from the parsed file. Note that the names for a parsed file contains only names for parts of decls, since we don't know references within terms before typechecking due to TDNR. This should be fine though, since those references will all be kept in the ABT as symbols anyways.

See UF.toNames and UF.typecheckedToNames for more info.

analyseNotes :: (Foldable f, MonadIO m) => Codebase IO Symbol Ann -> Uri -> PrettyPrintEnv -> String -> f (Note Symbol Ann) -> m ([Diagnostic], [RangedCodeAction])