Unison.LSP.Queries

Description

Rewrites of some codebase queries, but which check the scratch file for info first.

Documentation

refInTerm :: Term v a -> Maybe LabeledDependency

Returns the reference a given term node refers to, if any.

refInType :: Type v a -> Maybe TypeReference

findSmallestEnclosingNode :: Pos -> Term Symbol Ann -> Maybe (Either (Term Symbol Ann) (Type Symbol Ann))

Find the the node in a term which contains the specified position, but none of its children contain that position.

findSmallestEnclosingType :: Pos -> Type Symbol Ann -> Maybe (Type Symbol Ann)

Find the the node in a type which contains the specified position, but none of its children contain that position. This is helpful for finding the specific type reference of a given argument within a type arrow that a position references.

refInDecl :: Pos -> Decl Symbol Ann -> Maybe TypeReference

Returns the type reference the given position applies to within a Decl, if any.

I.e. if the cursor is over a type reference within a constructor signature or ability request signature, that type reference will be returned.