Unison.LSP.Types

Documentation

newtype Lsp a

A custom LSP monad wrapper so we can provide our own environment.

Constructors

Lsp

Fields

runLspM :: ReaderT Env (LspM Config) a

logInfo :: Text -> Lsp ()

Log an info message to the client's LSP log.

logError :: Text -> Lsp ()

Log an error message to the client's LSP log, this will be shown to the user in most LSP implementations.

data Env

Environment for the Lsp monad.

Constructors

Env

Fields

lspContext :: LanguageContextEnv Config

codebase :: Codebase IO Symbol Ann

parseNamesCache :: IO NamesWithHistory

ppeCache :: IO PrettyPrintEnvDecl

currentPathCache :: IO Absolute

vfsVar :: MVar VFS

runtime :: Runtime Symbol

checkedFilesVar :: TVar (Map Uri FileAnalysis)

dirtyFilesVar :: TVar (Set Uri)

cancellationMapVar :: TVar (Map SomeLspId (IO ()))

scope :: Scope

type FileVersion = Int32

A monotonically increasing file version tracked by the lsp client.

type LexedSource = (Text, [Token Lexeme])

data FileAnalysis

Constructors

FileAnalysis

Fields

fileUri :: Uri

fileVersion :: FileVersion

lexedSource :: LexedSource

parsedFile :: Maybe (UnisonFile Symbol Ann)

typecheckedFile :: Maybe (TypecheckedUnisonFile Symbol Ann)

notes :: Seq (Note Symbol Ann)

diagnostics :: IntervalMap Position [Diagnostic]

codeActions :: IntervalMap Position [CodeAction]

getCurrentPath :: Lsp Absolute

globalPPE :: Lsp PrettyPrintEnvDecl

getParseNames :: Lsp NamesWithHistory

data Config

Constructors

Config

lspBackend :: Backend IO a -> Lsp (Either BackendError a)

Lift a backend computation into the Lsp monad.

sendNotification :: forall (m :: Method 'FromServer 'Notification). Message m ~ NotificationMessage m => NotificationMessage m -> Lsp ()

data RangedCodeAction

Constructors

RangedCodeAction

Fields

_codeActionRanges :: [Range]

_codeAction :: CodeAction

rangedCodeAction :: Text -> [Diagnostic] -> [Range] -> RangedCodeAction

includeEdits :: Uri -> Text -> [Range] -> RangedCodeAction -> RangedCodeAction

Provided ranges must not intersect.