Unison.LSP.VFS.CompletionPrefix

Description

This code was removed from the lsp package in version 2.4.0.0. It’s vendored here until it’s replaced with a Unison- specific version.

Documentation

data PosPrefixInfo

Describes the line at the current cursor position

Constructors

PosPrefixInfo

Fields

fullLine :: !Text

The full contents of the line the cursor is at

prefixModule :: !Text

If any, the module name that was typed right before the cursor position. For example, if the user has typed "Data.Maybe.from", then this property will be Data.Maybe

prefixText :: !Text

The word right before the cursor position, after removing the module part. For example if the user has typed "Data.Maybe.from", then this property will be "from"

cursorPos :: !Position

The cursor position

getCompletionPrefix :: Monad m => Position -> VirtualFile -> m (Maybe PosPrefixInfo)