Unison.Share.Codeserver

Documentation

isCustomCodeserver :: CodeserverURI -> Bool

defaultCodeserver :: CodeserverURI

This is the URI where the share API is based.

resolveCodeserver :: ShareCodeserver -> CodeserverURI

data CodeserverURI

This type is expanded out into all of its fields because we require certain pieces which are optional in a URI, and also to make it more typesafe to eventually convert into a BaseURL for servant clients.

Constructors

CodeserverURI

Fields

codeserverScheme :: Scheme

codeserverUserInfo :: String

codeserverRegName :: String

codeserverPort :: Maybe Int

codeserverPath :: [String]