Unison.Share.ExpectedHashMismatches

Description

Unfortunately, due to bugs in previous versions of UCM there are some entities whose hashes are not what they should be, but have propagated widely into end-user's codebases. This is a map from the nominal hash to the actual hash of the component of components which are known to have widely available incorrect hashes.

Keeping this table allows us to allow UCM to validate the hashes of new terms which are received from remote sources, but allows a bit of lee-way for these known exceptions.

While we certainly don't want to make this a common occurence, it's safe as long as we are specific about both the provided and actual hashes.

Documentation

expectedComponentHashMismatches :: Map Hash32 Hash32

expectedCausalHashMismatches :: Map Hash32 Hash32