Unison.Share.Sync

High-level API

Get causal hash by path

getCausalHashByPath

Arguments

:: AuthenticatedHttpClient
The HTTP client to use for Unison Share requests.

-> BaseUrl
The Unison Share URL.

-> Path

-> IO (Either GetCausalHashByPathError (Maybe HashJWT))

Get the causal hash of a path hosted on Unison Share.

data GetCausalHashByPathError

An error occurred when getting causal hash by path.

Constructors

GetCausalHashByPathErrorNoReadPermission Path
The user does not have permission to read this path.

Push

checkAndSetPush

Arguments

:: AuthenticatedHttpClient
The HTTP client to use for Unison Share requests.

-> BaseUrl
The Unison Share URL.

-> Connection
SQLite connection, for reading entities to push.

-> Path
The repo+path to push to.

-> Maybe Hash32
The hash that we expect this repo+path to be at on Unison Share. If not, we'll get back a hash mismatch error. This prevents accidentally pushing over data that we didn't know was there.

-> CausalHash
The hash of our local causal to push.

-> (Int -> Int -> IO ())
Callback that is given the total number of entities uploaded, and the number of outstanding entities to upload.

-> IO (Either (SyncError CheckAndSetPushError) ())

Perform a check-and-set push (initially of just a causal hash, but ultimately all of its dependencies that the server is missing, too) to Unison Share.

This flavor of push takes the expected state of the server, and the desired state we want to set; if our expectation is off, we won't proceed with the push.

data CheckAndSetPushError

Error used by the client when pushing code to Unison Share.

Constructors

CheckAndSetPushErrorHashMismatch HashMismatch

CheckAndSetPushErrorNoWritePermission Path

CheckAndSetPushErrorServerMissingDependencies (NESet Hash32)

fastForwardPush

Arguments

:: AuthenticatedHttpClient
The HTTP client to use for Unison Share requests.

-> BaseUrl
The Unison Share URL.

-> Connection
SQLite connection, for reading entities to push.

-> Path
The repo+path to push to.

-> CausalHash
The hash of our local causal to push.

-> (Int -> Int -> IO ())
Callback that is given the total number of entities uploaded, and the number of outstanding entities to upload.

-> IO (Either (SyncError FastForwardPushError) ())

Perform a fast-forward push (initially of just a causal hash, but ultimately all of its dependencies that the server is missing, too) to Unison Share.

This flavor of push provides the server with a chain of causal hashes leading from its current state to our desired state.

data FastForwardPushError

An error occurred while fast-forward pushing code to Unison Share.

Constructors

FastForwardPushErrorNoHistory Path

FastForwardPushErrorNoReadPermission Path

FastForwardPushErrorNotFastForward Path

FastForwardPushErrorNoWritePermission Path

FastForwardPushErrorServerMissingDependencies (NESet Hash32)

FastForwardPushInvalidParentage Hash32 Hash32

Pull

pull

Arguments

:: AuthenticatedHttpClient
The HTTP client to use for Unison Share requests.

-> BaseUrl
The Unison Share URL.

-> Connection
SQLite connection, for writing entities we pull.

-> Path
The repo+path to pull from.

-> (Int -> IO ())
Callback that is given the total number of entities downloaded.

-> IO (Either (SyncError PullError) CausalHash)

data PullError

An error occurred while pulling code from Unison Share.

Constructors

PullErrorGetCausalHashByPath GetCausalHashByPathError
An error occurred while resolving a repo+path to a causal hash.

PullErrorNoHistoryAtPath Path