Unison.Share.Sync.Types

Description

Types used by the UCM client during sync.

Documentation

data CheckAndSetPushError

Error used by the client when pushing code to Unison Share.

Constructors

CheckAndSetPushErrorHashMismatch HashMismatch

CheckAndSetPushErrorNoWritePermission Path

CheckAndSetPushErrorServerMissingDependencies (NESet Hash32)

data FastForwardPushError

An error occurred while fast-forward pushing code to Unison Share.

Constructors

FastForwardPushErrorNoHistory Path

FastForwardPushErrorNoReadPermission Path

FastForwardPushErrorNotFastForward Path

FastForwardPushErrorNoWritePermission Path

FastForwardPushErrorServerMissingDependencies (NESet Hash32)

FastForwardPushInvalidParentage Hash32 Hash32

data PullError

An error occurred while pulling code from Unison Share.

Constructors

PullErrorGetCausalHashByPath GetCausalHashByPathError
An error occurred while resolving a repo+path to a causal hash.

PullErrorNoHistoryAtPath Path

data GetCausalHashByPathError

An error occurred when getting causal hash by path.

Constructors

GetCausalHashByPathErrorNoReadPermission Path
The user does not have permission to read this path.

data CodeserverTransportError

Generic Codeserver transport errors

Constructors

Unauthenticated BaseUrl

PermissionDenied Text

UnreachableCodeserver BaseUrl

InvalidResponse Response

RateLimitExceeded

InternalServerError

Timeout

data SyncError e

Constructors

TransportError CodeserverTransportError

SyncError e