Unison.Share.SyncV2

Documentation

syncFromFile

Arguments

:: Bool
Whether to validate entities as they're imported.

-> FilePath
Location of the sync-file

-> Cli (Either (SyncError PullError) CausalHash)

syncToFile

Arguments

:: Codebase IO v a

-> CausalHash
Root hash to sync

-> Maybe BranchRef
Optional name of the branch begin synced

-> FilePath
Location of the sync-file

-> IO (Either SyncErr ())

Sync a given causal hash and its dependencies to a sync-file.

syncFromCodebase

Arguments

:: Bool

-> Connection
The codebase to sync from.

-> Codebase IO v a

-> CausalHash
The hash to sync.

-> IO (Either (SyncError PullError) ())

syncFromCodeserver

Arguments

:: Bool

-> BaseUrl
The Unison Share URL.

-> BranchRef
The branch to download from.

-> HashJWT
The hash to download.

-> (Int -> IO ())
Callback that's given a number of entities we just downloaded.

-> Cli (Either (SyncError PullError) ())