Unison.Cli.DownloadUtils

Description

Utility functions for downloading remote entities and storing them locally in SQLite.

These are shared by commands like pull and clone.

Documentation

downloadProjectBranchFromShare :: HasCallStack => IncludeSquashedHead -> RemoteProjectBranch -> Cli (Either ShareError CausalHash)

Download a project/branch from Share.

downloadLooseCodeFromShare :: ReadShareLooseCode -> Cli (Either ShareError CausalHash)

Download loose code from Share.

data SyncVersion

Constructors

SyncV1

SyncV2