Unison.Cli.MonadUtils

Description

This module contains miscellaneous helper utils for rote actions in the Cli monad, like resolving a relative path to an absolute path, per the current path.

Paths

getCurrentPath :: Cli Absolute

Get the current path relative to the current project.

getCurrentProjectName :: Cli ProjectName

getCurrentProjectBranchName :: Cli ProjectBranchName

getCurrentProjectPath :: Cli ProjectPath

resolvePath :: Path -> Cli ProjectPath

Resolve a Path (interpreted as relative) to a Path.Absolute, per the current path.

resolvePath' :: Path' -> Cli ProjectPath

Resolve a Path' to a Path.Absolute, per the current path.

resolvePath'ToAbsolute :: Path' -> Cli Absolute

resolveSplit' :: (Path', a) -> Cli (ProjectPath, a)

Resolve a path split, per the current path.

Project and branch resolution

getCurrentProjectAndBranch :: Cli (ProjectAndBranch Project ProjectBranch)

getCurrentProjectBranch :: Cli ProjectBranch

Branches

Resolving branch identifiers

resolveAbsBranchId :: AbsBranchId -> Cli (Branch IO)

Resolve an AbsBranchId to the corresponding Branch IO, or fail if no such branch hash is found. (Non-existent branches by path are OK - the empty branch will be returned).

resolveAbsBranchIdV2 :: (forall void. Output -> Transaction void) -> ProjectAndBranch Project ProjectBranch -> AbsBranchId -> Transaction (Branch Transaction)

V2 version of resolveAbsBranchId2.

resolveBranchId :: BranchId -> Cli (Branch IO)

Resolve a BranchId to the corresponding Branch IO, or fail if no such branch hash is found. (Non-existent branches by path are OK - the empty branch will be returned).

resolveBranchIdToAbsBranchId :: BranchId -> Cli AbsBranchId

Resolve a BranchId to an AbsBranchId.

resolveShortCausalHash :: ShortCausalHash -> Cli (Branch IO)

Resolve a ShortCausalHash to the corresponding Branch IO, or fail if no such branch hash is found.

Getting/setting branches

getCurrentProjectRoot :: Cli (Branch IO)

Get the root branch.

getCurrentProjectRoot0 :: Cli (Branch0 IO)

Get the root branch0.

getCurrentBranch :: Cli (Branch IO)

Get the current branch.

getCurrentBranch0 :: Cli (Branch0 IO)

Get the current branch0.

getProjectBranchRoot :: ProjectBranch -> Cli (Branch IO)

getBranchFromProjectPath :: ProjectPath -> Cli (Branch IO)

Get the branch at an absolute path from the project root.

getBranch0FromProjectPath :: ProjectPath -> Cli (Branch0 IO)

Get the branch0 at an absolute path.

getMaybeBranchFromProjectPath :: ProjectPath -> Cli (Maybe (Branch IO))

Get the maybe-branch at an absolute path.

getMaybeBranch0FromProjectPath :: ProjectPath -> Cli (Maybe (Branch0 IO))

Get the maybe-branch0 at an absolute path.

expectBranchAtPath :: Path -> Cli (Branch IO)

Get the branch at a relative path, or return early if there's no such branch.

expectBranchAtPath' :: Path' -> Cli (Branch IO)

Get the branch at an absolute or relative path, or return early if there's no such branch.

expectBranch0AtPath :: Path -> Cli (Branch0 IO)

Get the branch0 at a relative path, or return early if there's no such branch.

expectBranch0AtPath' :: Path' -> Cli (Branch0 IO)

Get the branch0 at an absolute or relative path, or return early if there's no such branch.

assertNoBranchAtPath' :: Path' -> Cli ()

Assert that there's "no branch" at an absolute or relative path, or return early if there is one, where "no branch" means either there's actually no branch, or there is a branch whose head is empty (i.e. it may have a history, but no current terms/types etc).

branchExistsAtPath' :: Path' -> Cli Bool

Check if there's a branch at an absolute or relative path

"no branch" means either there's actually no branch, or there is a branch whose head is empty (i.e. it may have a history, but no current terms/types etc).

Updating branches

stepAt' :: Text -> (ProjectPath, Branch0 IO -> Cli (Branch0 IO)) -> Cli Bool

stepAt :: Text -> (ProjectPath, Branch0 IO -> Branch0 IO) -> Cli ()

stepAtM :: Text -> (ProjectPath, Branch0 IO -> IO (Branch0 IO)) -> Cli ()

stepManyAt :: ProjectBranch -> Text -> [(Absolute, Branch0 IO -> Branch0 IO)] -> Cli ()

stepManyAtM :: ProjectBranch -> Text -> [(Absolute, Branch0 IO -> IO (Branch0 IO))] -> Cli ()

updateProjectBranchRoot :: ProjectBranch -> Text -> (Branch IO -> Cli (Branch IO, r)) -> Cli r

updateProjectBranchRoot_ :: ProjectBranch -> Text -> (Branch IO -> Branch IO) -> Cli ()

updateAtM :: Text -> ProjectPath -> (Branch IO -> Cli (Branch IO)) -> Cli Bool

Update a branch at the given path, returning True if an update occurred and false otherwise

updateAt :: Text -> ProjectPath -> (Branch IO -> Branch IO) -> Cli Bool

Update a branch at the given path, returning True if an update occurred and false otherwise

updateAndStepAt :: (Foldable f, Foldable g, Functor g) => Text -> ProjectBranch -> f (Absolute, Branch IO -> Branch IO) -> g (Absolute, Branch0 IO -> Branch0 IO) -> Cli ()

Terms

getTermsAt :: (ProjectPath, HQSegment) -> Cli (Set Referent)

Types

getTypesAt :: (ProjectPath, HQSegment) -> Cli (Set TypeReference)

Patches

Default patch

defaultPatchPath :: Split'

The default patch path.

Getting patches

getPatchAt :: Split' -> Cli Patch

Get the patch at a path, or the empty patch if there's no such patch.

Latest touched Unison file

getLatestFile :: Cli (Maybe (FilePath, Bool))

getLatestParsedFile :: Cli (Maybe (UnisonFile Symbol Ann))

Get the latest parsed unison file.

getNamesFromLatestFile :: Cli Names

Gets the names from the latest typechecked unison file, or latest parsed file if it didn't typecheck.

getTermFromLatestParsedFile :: HashQualified Name -> Cli (Maybe (Term Symbol Ann))

Returns a parsed term (potentially with free variables) from the latest file. This term will refer to other terms in the file by vars, not by hash. Used to implement rewriting and other refactorings on the current file.

expectLatestFile :: Cli (FilePath, Bool)

expectLatestParsedFile :: Cli (UnisonFile Symbol Ann)

getLatestTypecheckedFile :: Cli (Maybe (TypecheckedUnisonFile Symbol Ann))

Get the latest typechecked unison file.

expectLatestTypecheckedFile :: Cli (TypecheckedUnisonFile Symbol Ann)

Get the latest typechecked unison file, or return early if there isn't one.

Parsing env

makeParsingEnv :: ProjectPath -> Names -> Cli (ParsingEnv Transaction)