Unison.Cli.NamesUtils

Description

Utilities that have to do with constructing names objects.

Documentation

currentNames :: Cli Names

Produce a Names object which contains names for the current branch.

currentProjectRootNames :: Cli Names

projectBranchNames :: ProjectBranch -> Cli Names