Unison.Cli.ProjectUtils

Description

Project-related utilities.

Project/path helpers

getCurrentProject :: Cli (Maybe Project)

Get the current project that a user is on.

expectCurrentProject :: Cli Project

Like getCurrentProject, but fails with a message if the user is not on a project branch.

expectCurrentProjectIds :: Cli (ProjectAndBranch ProjectId ProjectBranchId)

Like getCurrentProjectIds, but fails with a message if the user is not on a project branch.

getCurrentProjectIds :: Cli (Maybe (ProjectAndBranch ProjectId ProjectBranchId))

Get the current project ids that a user is on.

getCurrentProjectBranch :: Cli (Maybe (ProjectAndBranch Project ProjectBranch, Path))

Get the current project+branch+branch path that a user is on.

getProjectBranchForPath :: Absolute -> Cli (Maybe (ProjectAndBranch Project ProjectBranch, Path))

expectCurrentProjectBranch :: Cli (ProjectAndBranch Project ProjectBranch, Path)

Like getCurrentProjectBranch, but fails with a message if the user is not on a project branch.

projectPath :: ProjectId -> Absolute

Get the path that a project is stored at. Users aren't supposed to go here.

>>> projectPath "ABCD"
.__projects._ABCD

projectBranchesPath :: ProjectId -> Absolute

Get the path that a project's branches are stored at. Users aren't supposed to go here.

>>> projectBranchesPath "ABCD"
.__projects._ABCD.branches

projectBranchPath :: ProjectAndBranch ProjectId ProjectBranchId -> Absolute

Get the path that a branch is stored at. Users aren't supposed to go here.

>>> projectBranchPath ProjectAndBranch { project = "ABCD", branch = "DEFG" }
.__projects._ABCD.branches._DEFG

projectBranchSegment :: ProjectBranchId -> NameSegment

Get the name segment that a branch is stored at.

>>> projectBranchSegment "DEFG"
"_DEFG"

projectBranchPathPrism :: Prism' Absolute (ProjectAndBranch ProjectId ProjectBranchId, Path)

The prism between paths like

.__projects._XX_XX.branches._YY_YY.foo.bar

and the (project id, branch id, path) triple

(XX-XX, YY-YY, foo.bar)

resolveBranchRelativePath :: BranchRelativePath -> Cli ResolvedBranchRelativePath

branchRelativePathToAbsolute :: BranchRelativePath -> Cli Absolute

Name hydration

hydrateNames :: These ProjectName ProjectBranchName -> Cli (ProjectAndBranch ProjectName ProjectBranchName)

Loading local project info

expectProjectAndBranchByIds :: ProjectAndBranch ProjectId ProjectBranchId -> Transaction (ProjectAndBranch Project ProjectBranch)

getProjectAndBranchByTheseNames :: These ProjectName ProjectBranchName -> Cli (Maybe (ProjectAndBranch Project ProjectBranch))

expectProjectAndBranchByTheseNames :: These ProjectName ProjectBranchName -> Cli (ProjectAndBranch Project ProjectBranch)

expectLooseCodeOrProjectBranch :: These Path' (ProjectAndBranch (Maybe ProjectName) ProjectBranchName) -> Cli (Either Path' (ProjectAndBranch Project ProjectBranch))

Expect/resolve a possibly-ambiguous "loose code or project", with the following rules:

If we have an unambiguous `branch` or `projectbranch`, look up in the database.

If we have an unambiguous `loose.code.path`, just return it.

If we have an ambiguous foo, *because we do not currently have an unambiguous syntax for relative paths*, we elect to treat it as a loose code path (because `/branch` can be selected with a leading forward slash).

Loading remote project info

expectRemoteProjectByName :: ProjectName -> Cli RemoteProject

expectRemoteProjectBranchById :: IncludeSquashedHead -> ProjectAndBranch (RemoteProjectId, ProjectName) (RemoteProjectBranchId, ProjectBranchName) -> Cli RemoteProjectBranch

loadRemoteProjectBranchByName :: IncludeSquashedHead -> ProjectAndBranch RemoteProjectId ProjectBranchName -> Cli (Maybe RemoteProjectBranch)

expectRemoteProjectBranchByName :: IncludeSquashedHead -> ProjectAndBranch (RemoteProjectId, ProjectName) ProjectBranchName -> Cli RemoteProjectBranch

loadRemoteProjectBranchByNames :: IncludeSquashedHead -> ProjectAndBranch ProjectName ProjectBranchName -> Cli (Maybe RemoteProjectBranch)

expectRemoteProjectBranchByNames :: IncludeSquashedHead -> ProjectAndBranch ProjectName ProjectBranchName -> Cli RemoteProjectBranch

expectRemoteProjectBranchByTheseNames :: IncludeSquashedHead -> These ProjectName ProjectBranchName -> Cli RemoteProjectBranch