Unison.Codebase.Editor.HandleInput.Branch

Description

branch input handler

Documentation

handleBranch :: BranchSourceI -> ProjectAndBranch (Maybe ProjectName) ProjectBranchName -> Cli ()

Create a new project branch from an existing project branch or namespace.

data CreateFrom

Constructors

CreateFrom'Branch (ProjectAndBranch Project ProjectBranch)

CreateFrom'LooseCode Absolute

CreateFrom'Nothingness

doCreateBranch :: CreateFrom -> Project -> ProjectBranchName -> Text -> Cli ProjectBranchId

doCreateBranch createFrom project branch description:

Creates a new branch row for branch in project project (failing if branch already exists in project)

Puts the branch contents from createFrom in the root namespace., using description for the reflog.

cds to the new branch in the root namespace.

This bit of functionality is factored out from the main handleBranch handler because it is also called by the release.draft command, which essentially just creates a branch, but with some different output for the user.

Returns the branch id of the newly-created branch.