Unison.Codebase.Editor.HandleInput.Names

Documentation

handleNames :: Bool -> (RawQuery, ErrorMessageOrName) -> Cli ()

Handles a single NamesI input query returning terms that match a given name.

Parameters:

global :: Bool

* If True, search all projects and branches.

* If False, search only the current branch.

query :: (RawQuery, ErrorMessageOrName)

* The first member is the raw nameQuery being handled.

* The second member is the parsed nameQuery that is either an error message to be printed or a name that can be looked up in the codebase.