Unison.Codebase.Editor.HandleInput.ShowDefinition

Documentation

handleShowDefinition :: OutputLocation -> ShowDefinitionScope -> NonEmpty (HashQualified Name) -> Cli ()

Handle a ShowDefinitionI input command, i.e. view or edit.

showDefinitions :: OutputLocation -> PrettyPrintEnvDecl -> Map Reference (DisplayObject (Type Symbol Ann) (Term Symbol Ann)) -> Map Reference (DisplayObject () (Decl Symbol Ann)) -> [HashQualified Name] -> Cli ()

Show the provided definitions to console or scratch file. The caller is responsible for ensuring that the definitions include cycles if that's the desired behavior.