Unison.Codebase.Editor.HandleInput.Update2

Description

update input handler.

Documentation

handleUpdate2 :: Cli ()

Misc helpers to be organized later

typecheckedUnisonFileToBranchUpdates

Arguments

:: (forall void. Output -> Transaction void)

-> (Name -> Either Output (Maybe [Name]))
Returns Nothing if the decl isn't in namesExcludingLibdeps, in which case we know the decl is new and do not need to generate delete actions for it.

-> TypecheckedUnisonFile Symbol Ann

-> Transaction [(Path, Branch0 m -> Branch0 m)]