Unison.Codebase.Editor.Input

Documentation

data Input

Constructors

ForkLocalBranchI (Either ShortCausalHash BranchRelativePath) BranchRelativePath

MergeLocalBranchI BranchRelativePath (Maybe BranchRelativePath) MergeMode

PreviewMergeLocalBranchI BranchRelativePath (Maybe BranchRelativePath)

DiffNamespaceI BranchId2 BranchId2

PullI !PullSourceTarget !PullMode

PushRemoteBranchI PushRemoteBranchInput

ResetI BranchId2 (Maybe UnresolvedProjectBranch)

CreateMessage (Pretty ColorText)

SwitchBranchI Path'

UpI

PopBranchI

NamesI IsGlobal (HashQualified Name)

AliasTermI !Bool HashOrHQSplit' Split'

AliasTypeI !Bool HashOrHQSplit' Split'

AliasManyI [HQSplit] Path'

MoveAllI Path' Path'

MoveTermI HQSplit' Split'

MoveTypeI HQSplit' Split'

MoveBranchI Path' Path'

DeleteI DeleteTarget

LoadI (Maybe FilePath)

ClearI

AddI (Set Name)

PreviewAddI (Set Name)

UpdateI OptionalPatch (Set Name)

Update2I

PreviewUpdateI (Set Name)

TodoI

UndoI

HistoryI (Maybe Int) (Maybe Int) BranchId

ExecuteI (HashQualified Name) [String]

SaveExecuteResultI Name

IOTestI (HashQualified Name)

IOTestAllI

MakeStandaloneI String (HashQualified Name)

ExecuteSchemeI (HashQualified Name) [String]

CompileSchemeI Bool Text (HashQualified Name)

TestI TestInput

CreateAuthorI NameSegment Text

DisplayI OutputLocation (NonEmpty (HashQualified Name))

DocsI (NonEmpty Name)

FindI Bool FindScope [String]

FindShallowI Path'

StructuredFindI FindScope (HashQualified Name)

StructuredFindReplaceI (HashQualified Name)

TextFindI Bool [String]

ShowDefinitionI OutputLocation ShowDefinitionScope (NonEmpty (HashQualified Name))

ShowRootReflogI

ShowGlobalReflogI

ShowProjectReflogI (Maybe ProjectName)

ShowProjectBranchReflogI (Maybe (ProjectAndBranch (Maybe ProjectName) ProjectBranchName))

UpdateBuiltinsI

MergeBuiltinsI (Maybe Path)

MergeIOBuiltinsI (Maybe Path)

ListDependenciesI (HashQualified Name)

ListDependentsI (HashQualified Name)

NamespaceDependenciesI (Maybe Path')
List all external dependencies of a given namespace, or the current namespace if no path is provided.

DebugTabCompletionI [String]

DebugLSPNameCompletionI Text

DebugFuzzyOptionsI String [String]

DebugFormatI

DebugNumberedArgsI

DebugTypecheckedUnisonFileI

DebugDumpNamespacesI

DebugDumpNamespaceSimpleI

DebugTermI Bool (HashQualified Name)

DebugTypeI (HashQualified Name)

DebugLSPFoldRangesI

DebugClearWatchI

DebugDoctorI

DebugNameDiffI ShortCausalHash ShortCausalHash

QuitI

ApiI

UiI Path'

DocToMarkdownI Name

DocsToHtmlI BranchRelativePath FilePath

AuthLoginI

VersionI

ProjectCreateI Bool (Maybe ProjectName)

ProjectRenameI ProjectName

ProjectSwitchI ProjectAndBranchNames

ProjectsI

BranchI BranchSourceI (ProjectAndBranch (Maybe ProjectName) ProjectBranchName)

BranchRenameI ProjectBranchName

BranchesI (Maybe ProjectName)

CloneI ProjectAndBranchNames (Maybe ProjectAndBranchNames)

ReleaseDraftI Semver

UpgradeI !NameSegment !NameSegment

EditNamespaceI [Path]

MergeI (ProjectAndBranch (Maybe ProjectName) ProjectBranchName)

LibInstallI !Bool !(ProjectAndBranch ProjectName (Maybe ProjectBranchNameOrLatestRelease))

UpgradeCommitI

MergeCommitI

DebugSynhashTermI !Name

data BranchSourceI

The source of a branch command: what to make the new branch from.

Constructors

BranchSourceI'CurrentContext
Create a branch from the current context

BranchSourceI'Empty
Create an empty branch

BranchSourceI'UnresolvedProjectBranch UnresolvedProjectBranch
Create a branch from this other branch

data PullSourceTarget

Pull source and target: either neither is specified, or only a source, or both.

Constructors

PullSourceTarget0

PullSourceTarget1 (ReadRemoteNamespace (These ProjectName ProjectBranchNameOrLatestRelease))

PullSourceTarget2 (ReadRemoteNamespace (These ProjectName ProjectBranchNameOrLatestRelease)) (ProjectAndBranch (Maybe ProjectName) ProjectBranchName)

data PushRemoteBranchInput

Constructors

PushRemoteBranchInput

Fields

sourceTarget :: PushSourceTarget

pushBehavior :: PushBehavior

data PushSourceTarget

Push source and target: either neither is specified, or only a target, or both.

Constructors

PushSourceTarget0

PushSourceTarget1 (These ProjectName ProjectBranchName)

PushSourceTarget2 PushSource (These ProjectName ProjectBranchName)

data PushSource

Constructors

ProjySource (These ProjectName ProjectBranchName)

data TestInput

Constructors

TestInput

Fields

includeLibNamespace :: Bool

Should we run tests in the lib namespace?

path :: Path

Relative path to run the tests in. Ignore if $sel:includeLibNamespace:TestInput is True - that means test everything.

showFailures :: Bool

showSuccesses :: Bool

data Event

Constructors

UnisonFileChanged SourceName Source

data OutputLocation

Constructors

ConsoleLocation

LatestFileLocation RelativeToFold

FileLocation FilePath RelativeToFold

data RelativeToFold

Above a new fold, or within the topmost fold?

Constructors

AboveFold

WithinFold

type PatchPath = Split'

data BranchIdG p

Constructors

BranchAtSCH ShortCausalHash

BranchAtPath p

BranchAtProjectPath ProjectPath

type BranchId = BranchIdG Path'

type BranchId2 = Either ShortCausalHash BranchRelativePath

type AbsBranchId = BranchIdG Absolute

type UnresolvedProjectBranch = ProjectAndBranch (Maybe ProjectName) ProjectBranchName

An unambiguous project branch name, use the current project name if not provided.

parseBranchId :: String -> Either Text BranchId

parseBranchId2 :: String -> Either (Pretty ColorText) (Either ShortCausalHash BranchRelativePath)

parseShortCausalHash :: String -> Either String ShortCausalHash

type HashOrHQSplit' = Either ShortHash HQSplit'

data Insistence

Should we force the operation or not?

Constructors

Force

Try

data PullMode

Constructors

PullWithHistory

PullWithoutHistory

data OptionalPatch

Constructors

NoPatch

DefaultPatch

UsePatch PatchPath

data FindScope

Constructors

FindLocal Path'

FindLocalAndDeps Path'

FindGlobal

data ShowDefinitionScope

Constructors

ShowDefinitionLocal

ShowDefinitionGlobal

type IsGlobal = Bool

data DeleteOutput

Constructors

DeleteOutput'Diff

DeleteOutput'NoDiff

data DeleteTarget

Constructors

DeleteTarget'TermOrType DeleteOutput [HQSplit']

DeleteTarget'Term DeleteOutput [HQSplit']

DeleteTarget'Type DeleteOutput [HQSplit']

DeleteTarget'Namespace Insistence (Maybe Split)

DeleteTarget'ProjectBranch (ProjectAndBranch (Maybe ProjectName) ProjectBranchName)

DeleteTarget'Project ProjectName