Unison.Codebase.Editor.Output

Documentation

data Output

Constructors

Success

NoUnisonFile

PrintMessage (Pretty ColorText)

InvalidSourceName String

SourceLoadFailed String

NoMainFunction String PrettyPrintEnv [Type Symbol Ann]

BadMainFunction
Function found, but has improper type Note: the constructor name is misleading here; we weren't necessarily looking for a "main".

Fields

String

what we were trying to do (e.g. "run", "io.test")

String

name of function

(Type Symbol Ann)

bad type of function

PrettyPrintEnv

[Type Symbol Ann]

acceptable type(s) of function

BranchEmpty WhichBranchEmpty

BranchNotEmpty Path'

LoadPullRequest ReadRemoteNamespace ReadRemoteNamespace Path' Path' Path' Path'

CreatedNewBranch Absolute

BranchAlreadyExists Path'

FindNoLocalMatches

PatchAlreadyExists Split'

NoExactTypeMatches

TypeAlreadyExists Split' (Set Reference)

TypeParseError String (Err Symbol)

ParseResolutionFailures String [ResolutionFailure Symbol Ann]

TypeHasFreeVars (Type Symbol Ann)

TermAlreadyExists Split' (Set Referent)

LabeledReferenceAmbiguous Int (HashQualified Name) (Set LabeledDependency)

LabeledReferenceNotFound (HashQualified Name)

DeleteNameAmbiguous Int HQSplit' (Set Referent) (Set Reference)

TermAmbiguous (HashQualified Name) (Set Referent)

HashAmbiguous ShortHash (Set Referent)

BranchHashAmbiguous ShortCausalHash (Set ShortCausalHash)

BadNamespace String String

BranchNotFound Path'

EmptyPush Path'

NameNotFound HQSplit'

NamesNotFound [Name]

PatchNotFound Split'

TypeNotFound HQSplit'

TermNotFound HQSplit'

TypeNotFound' ShortHash

TermNotFound' ShortHash

TypeTermMismatch (HashQualified Name) (HashQualified Name)

NoLastRunResult

SaveTermNameConflict Name

SearchTermsNotFound [HashQualified Name]

SearchTermsNotFoundDetailed

Fields

Bool

True if we are searching for a term, False if we are searching for a type

[HashQualified Name]

Misses (search terms that returned no hits for terms or types)

[HashQualified Name]

Hits for types if we are searching for terms or terms if we are searching for types

DeleteBranchConfirmation [(Path', (Names, [SearchResult' Symbol Ann]))]

DeleteEverythingConfirmation

MoveRootBranchConfirmation

MovedOverExistingBranch Path'

DeletedEverything

ListNames IsGlobal Int [(Reference, [HashQualified Name])] [(Referent, [HashQualified Name])]

ListOfDefinitions FindScope PrettyPrintEnv ListDetailed [SearchResult' Symbol Ann]

ListOfLinks PrettyPrintEnv [(HashQualified Name, Reference, Maybe (Type Symbol Ann))]

ListShallow (IO PrettyPrintEnv) [ShallowListEntry Symbol Ann]

ListOfPatches (Set Name)

SlurpOutput Input PrettyPrintEnv SlurpResult

ParseErrors Text [Err Symbol]

TypeErrors Absolute Text PrettyPrintEnv [ErrorNote Symbol Ann]

CompilerBugs Text PrettyPrintEnv [CompilerBug Symbol Ann]

DisplayConflicts (Relation Name Referent) (Relation Name Reference)

EvaluationFailure Error

Evaluated SourceFileContents PrettyPrintEnv [(Symbol, Term Symbol ())] (Map Symbol (Ann, WatchKind, Term Symbol (), IsCacheHit))

RunResult PrettyPrintEnv (Term Symbol ())

Typechecked SourceName PrettyPrintEnv SlurpResult (TypecheckedUnisonFile Symbol Ann)

DisplayRendered (Maybe FilePath) (Pretty ColorText)

DisplayDefinitions DisplayDefinitionsOutput

TestIncrementalOutputStart PrettyPrintEnv (Int, Int) Reference (Term Symbol Ann)

TestIncrementalOutputEnd PrettyPrintEnv (Int, Int) Reference (Term Symbol Ann)

TestResults TestReportStats PrettyPrintEnv ShowSuccesses ShowFailures [(Reference, Text)] [(Reference, Text)]

CantUndo UndoFailureReason

BustedBuiltins (Set Reference) (Set Reference)

GitError GitError

ShareError ShareError

ViewOnShare WriteShareRemotePath

ConfiguredMetadataParseError Path' String (Pretty ColorText)

NoConfiguredRemoteMapping PushPull Absolute

ConfiguredRemoteMappingParseError PushPull Absolute Text String

MetadataMissingType PrettyPrintEnv Referent

TermMissingType Reference

MetadataAmbiguous (HashQualified Name) PrettyPrintEnv [Referent]

NothingToPatch PatchPath Path'

PatchNeedsToBeConflictFree

PatchInvolvesExternalDependents PrettyPrintEnv (Set Reference)

WarnIncomingRootBranch ShortCausalHash (Set ShortCausalHash)

StartOfCurrentPathHistory

ShowReflog [(Maybe UTCTime, ShortCausalHash, Text)]

PullAlreadyUpToDate ReadRemoteNamespace Path'

PullSuccessful ReadRemoteNamespace Path'

MergeOverEmpty Path'
Indicates a trivial merge where the destination was empty and was just replaced.

MergeAlreadyUpToDate Path' Path'

PreviewMergeAlreadyUpToDate Path' Path'

NoConflictsOrEdits
No conflicts or edits remain for the current patch.

NotImplemented

NoBranchWithHash ShortCausalHash

ListDependencies Int LabeledDependency [(Name, Reference)] (Set Reference)

ListDependents Int LabeledDependency [(Reference, Maybe Name)]
List dependents of a type or term.

ListNamespaceDependencies PrettyPrintEnv Absolute (Map LabeledDependency (Set Name))
List all direct dependencies which don't have any names in the current branch

DumpNumberedArgs NumberedArgs

DumpBitBooster CausalHash (Map CausalHash [CausalHash])

DumpUnisonFileHashes Int [(Name, Id)] [(Name, Id)] [(Name, Id)]

BadName String

DefaultMetadataNotification

CouldntLoadBranch CausalHash

HelpMessage InputPattern

NamespaceEmpty (NonEmpty AbsBranchId)

NoOp

RefusedToPush PushBehavior WriteRemotePath

GistCreated ReadRemoteNamespace
GistCreated repo means a causal was just published to repo.

InitiateAuthFlow URI
Directs the user to URI to begin an authorization flow.

UnknownCodeServer Text

CredentialFailureMsg CredentialFailure

PrintVersion Text

IntegrityCheck IntegrityResult

DisplayDebugNameDiff NameChanges

DisplayDebugCompletions [Completion]

ClearScreen

PulledEmptyBranch ReadRemoteNamespace

data DisplayDefinitionsOutput

Constructors

DisplayDefinitionsOutput

Fields

isTest :: TermReference -> Bool

outputFile :: Maybe FilePath

prettyPrintEnv :: PrettyPrintEnvDecl

terms :: Map Reference (DisplayObject (Type Symbol Ann) (Term Symbol Ann))

types :: Map Reference (DisplayObject () (Decl Symbol Ann))

data WhichBranchEmpty

A branch was empty. But how do we refer to that branch?

Constructors

WhichBranchEmptyHash ShortCausalHash

WhichBranchEmptyPath Path'

data NumberedOutput

Constructors

ShowDiffNamespace AbsBranchId AbsBranchId PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterUndo PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterDeleteDefinitions PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterDeleteBranch Absolute PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterModifyBranch Path' Absolute PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterMerge Path' Absolute PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterMergePropagate Path' Absolute Path' PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterMergePreview Path' Absolute PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterPull Path' Absolute PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterCreatePR ReadRemoteNamespace ReadRemoteNamespace PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterCreateAuthor NameSegment Path' Absolute PrettyPrintEnv (BranchDiffOutput Symbol Ann)

TodoOutput PrettyPrintEnvDecl (TodoOutput Symbol Ann)
Invariant: there's at least one conflict or edit in the TodoOutput.

CantDeleteDefinitions PrettyPrintEnvDecl (Map LabeledDependency (NESet LabeledDependency))
CantDeleteDefinitions ppe couldntDelete becauseTheseStillReferenceThem

CantDeleteNamespace PrettyPrintEnvDecl (Map LabeledDependency (NESet LabeledDependency))
CantDeleteNamespace ppe couldntDelete becauseTheseStillReferenceThem

DeletedDespiteDependents PrettyPrintEnvDecl (Map LabeledDependency (NESet LabeledDependency))
DeletedDespiteDependents ppe deletedThings thingsWhichNowHaveUnnamedReferences

History (Maybe Int) HashLength [(CausalHash, Diff)] HistoryTail
size limit, history , how the history ends

ListEdits Patch PrettyPrintEnv

type NumberedArgs = [String]

type ListDetailed = Bool

data HistoryTail

Constructors

EndOfLog CausalHash

MergeTail CausalHash [CausalHash]

PageEnd CausalHash Int

data TestReportStats

Constructors

CachedTests TotalCount CachedCount

NewlyComputed

data UndoFailureReason

Constructors

CantUndoPastStart

CantUndoPastMerge

data ShareError

Constructors

ShareErrorCheckAndSetPush CheckAndSetPushError

ShareErrorFastForwardPush FastForwardPushError

ShareErrorPull PullError

ShareErrorGetCausalHashByPath GetCausalHashByPathError

ShareErrorTransport CodeserverTransportError

isFailure :: Output -> Bool

isNumberedFailure :: NumberedOutput -> Bool