Unison.Codebase.Editor.Output

Documentation

data Output

Constructors

Success

NoUnisonFile

PrintMessage (Pretty ColorText)

InvalidSourceName String

SourceLoadFailed String

NoMainFunction (HashQualified Name) PrettyPrintEnv [Type Symbol Ann]

BadMainFunction
Function found, but has improper type Note: the constructor name is misleading here; we weren't necessarily looking for a "main".

Fields

Text

what we were trying to do (e.g. "run", "io.test")

(HashQualified Name)

name of function

(Type Symbol Ann)

bad type of function

PrettyPrintEnv

[Type Symbol Ann]

acceptable type(s) of function

BranchEmpty WhichBranchEmpty

LoadPullRequest (ReadRemoteNamespace Void) (ReadRemoteNamespace Void) Path' Path' Path' Path'

CreatedNewBranch Absolute

BranchAlreadyExists Path'

FindNoLocalMatches

NoExactTypeMatches

TypeAlreadyExists Split' (Set Reference)

TypeParseError String (Err Symbol)

ParseResolutionFailures String [ResolutionFailure Ann]

TypeHasFreeVars (Type Symbol Ann)

TermAlreadyExists Split' (Set Referent)

LabeledReferenceAmbiguous Int (HashQualified Name) (Set LabeledDependency)

LabeledReferenceNotFound (HashQualified Name)

DeleteNameAmbiguous Int HQSplit' (Set Referent) (Set Reference)

TermAmbiguous PrettyPrintEnv (HashQualified Name) (Set Referent)

HashAmbiguous ShortHash (Set Referent)

BranchHashAmbiguous ShortCausalHash (Set ShortCausalHash)

BadNamespace String String

BranchNotFound Path'

EmptyLooseCodePush Path'

EmptyProjectBranchPush (ProjectAndBranch ProjectName ProjectBranchName)

NameNotFound HQSplit'

NamesNotFound [Name]

TypeNotFound HQSplit'

TermNotFound HQSplit'

MoveNothingFound Path'

TypeNotFound' ShortHash

TermNotFound' ShortHash

NoLastRunResult

SaveTermNameConflict Name

SearchTermsNotFound [HashQualified Name]

SearchTermsNotFoundDetailed

Fields

Bool

True if we are searching for a term, False if we are searching for a type

[HashQualified Name]

Misses (search terms that returned no hits for terms or types)

[HashQualified Name]

Hits for types if we are searching for terms or terms if we are searching for types

DeleteBranchConfirmation [(Path', (Names, [SearchResult' Symbol Ann]))]

DeleteEverythingConfirmation

MoveRootBranchConfirmation

MovedOverExistingBranch Path'

DeletedEverything

ListNames Int [(Reference, [HashQualified Name])] [(Referent, [HashQualified Name])]

GlobalListNames (ProjectAndBranch ProjectName ProjectBranchName) Int [(Reference, [HashQualified Name])] [(Referent, [HashQualified Name])]

ListOfDefinitions FindScope PrettyPrintEnv ListDetailed [SearchResult' Symbol Ann]

ListShallow (IO PrettyPrintEnv) [ShallowListEntry Symbol Ann]

ListStructuredFind [HashQualified Name]

ListTextFind Bool [HashQualified Name]

GlobalFindBranchResults (ProjectAndBranch ProjectName ProjectBranchName) PrettyPrintEnv ListDetailed [SearchResult' Symbol Ann]

SlurpOutput Input PrettyPrintEnv SlurpResult

ParseErrors Text [Err Symbol]

TypeErrors Absolute Text PrettyPrintEnv [ErrorNote Symbol Ann]

CompilerBugs Text PrettyPrintEnv [CompilerBug Symbol Ann]

DisplayConflicts (Relation Name Referent) (Relation Name Reference)

EvaluationFailure Error

Evaluated SourceFileContents PrettyPrintEnv [(Symbol, Term Symbol ())] (Map Symbol (Ann, WatchKind, Term Symbol (), IsCacheHit))

RunResult PrettyPrintEnv (Term Symbol ())

LoadingFile SourceName

Typechecked SourceName PrettyPrintEnv SlurpResult (TypecheckedUnisonFile Symbol Ann)

DisplayRendered (Maybe FilePath) (Pretty ColorText)

DisplayDefinitions (Pretty ColorText)

LoadedDefinitionsToSourceFile FilePath Int

TestIncrementalOutputStart PrettyPrintEnv (Int, Int) TermReferenceId

TestIncrementalOutputEnd PrettyPrintEnv (Int, Int) TermReferenceId Bool

CantUndo UndoFailureReason

BustedBuiltins (Set Reference) (Set Reference)

ShareError ShareError

ViewOnShare (URI, ProjectName, ProjectBranchName)

NoConfiguredRemoteMapping PushPull Absolute

ConfiguredRemoteMappingParseError PushPull Absolute Text String

TermMissingType Reference

AboutToPropagatePatch

PatchNeedsToBeConflictFree

PatchInvolvesExternalDependents PrettyPrintEnv (Set Reference)

StartOfCurrentPathHistory

ShowReflog [(Maybe UTCTime, ShortCausalHash, Text)]

PullAlreadyUpToDate (ReadRemoteNamespace RemoteProjectBranch) (ProjectAndBranch Project ProjectBranch)

PullSuccessful (ReadRemoteNamespace RemoteProjectBranch) (ProjectAndBranch Project ProjectBranch)

AboutToMerge

MergeOverEmpty (ProjectAndBranch Project ProjectBranch)
Indicates a trivial merge where the destination was empty and was just replaced.

MergeAlreadyUpToDate BranchRelativePath BranchRelativePath

MergeAlreadyUpToDate2 !MergeSourceAndTarget

PreviewMergeAlreadyUpToDate ProjectPath ProjectPath

NotImplemented

NoBranchWithHash ShortCausalHash

ListDependencies PrettyPrintEnv (Set LabeledDependency) [HashQualified Name] [HashQualified Name]

ListDependents PrettyPrintEnv (Set LabeledDependency) [HashQualified Name] [HashQualified Name]
List dependents of a type or term.

DumpNumberedArgs HashLength NumberedArgs

DumpBitBooster CausalHash (Map CausalHash [CausalHash])

DumpUnisonFileHashes Int [(Name, Id)] [(Name, Id)] [(Name, Id)]

BadName Text

CouldntLoadBranch CausalHash

HelpMessage InputPattern

NamespaceEmpty (NonEmpty (Either ShortCausalHash ProjectPath))

NoOp

GistCreated (ReadRemoteNamespace Void)
GistCreated repo means a causal was just published to repo.

InitiateAuthFlow URI
Directs the user to URI to begin an authorization flow.

UnknownCodeServer Text

CredentialFailureMsg CredentialFailure

PrintVersion Text

IntegrityCheck IntegrityResult

DisplayDebugNameDiff NameChanges

DisplayDebugCompletions [Completion]

DisplayDebugLSPNameCompletions [(Text, Name, LabeledDependency)]

DebugDisplayFuzzyOptions Text [String]

DebugFuzzyOptionsNoResolver

DebugTerm Bool (Either Text (Term Symbol Ann))

DebugDecl (Either Text (Decl Symbol Ann)) (Maybe ConstructorId)

AnnotatedFoldRanges Text

ClearScreen

PulledEmptyBranch (ReadRemoteNamespace RemoteProjectBranch)

CreatedProject Bool ProjectName

CreatedProjectBranch CreatedProjectBranchFrom (ProjectAndBranch ProjectName ProjectBranchName)

CreatedRemoteProject URI (ProjectAndBranch ProjectName ProjectBranchName)

CreatedRemoteProjectBranch URI (ProjectAndBranch ProjectName ProjectBranchName)

RemoteProjectBranchIsUpToDate URI (ProjectAndBranch ProjectName ProjectBranchName)

InvalidProjectName Text

InvalidProjectBranchName Text

InvalidStructuredFindReplace (HashQualified Name)

InvalidStructuredFind (HashQualified Name)

ProjectNameAlreadyExists ProjectName

ProjectNameRequiresUserSlug ProjectName

ProjectAndBranchNameAlreadyExists (ProjectAndBranch ProjectName ProjectBranchName)

NotOnProjectBranch

NoAssociatedRemoteProject URI (ProjectAndBranch ProjectName ProjectBranchName)

NoAssociatedRemoteProjectBranch URI (ProjectAndBranch Project ProjectBranch)

LocalProjectDoesntExist ProjectName

LocalProjectBranchDoesntExist (ProjectAndBranch ProjectName ProjectBranchName)

LocalProjectNorProjectBranchExist ProjectName ProjectBranchName

RemoteProjectDoesntExist URI ProjectName

RemoteProjectBranchDoesntExist URI (ProjectAndBranch ProjectName ProjectBranchName)

RemoteProjectBranchDoesntExist'Push URI (ProjectAndBranch ProjectName ProjectBranchName)

RemoteProjectReleaseIsDeprecated URI (ProjectAndBranch ProjectName ProjectBranchName)

RemoteProjectPublishedReleaseCannotBeChanged URI (ProjectAndBranch ProjectName ProjectBranchName)

RemoteProjectBranchHeadMismatch URI (ProjectAndBranch ProjectName ProjectBranchName)

Unauthorized Text

ServantClientError ClientError

MarkdownOut Text

DownloadedEntities Int

UploadedEntities Int

NotImplementedYet Text

DraftingRelease ProjectBranchName Semver

CannotCreateReleaseBranchWithBranchCommand ProjectBranchName Semver

CalculatingDiff

AmbiguousCloneLocal
The local in a `clone remote local` is ambiguous

Fields

(ProjectAndBranch ProjectName ProjectBranchName)

Treating local as a project. We may know the branch name, if it was provided in remote.

(ProjectAndBranch ProjectName ProjectBranchName)

AmbiguousCloneRemote ProjectName (ProjectAndBranch ProjectName ProjectBranchName)
The remote in a `clone remote local` is ambiguous

ClonedProjectBranch (ProjectAndBranch ProjectName ProjectBranchName) (ProjectAndBranch ProjectName ProjectBranchName)

RenamedProject ProjectName ProjectName

OutputRewrittenFile FilePath [Symbol]

RenamedProjectBranch ProjectName ProjectBranchName ProjectBranchName

CantRenameBranchTo ProjectBranchName

FetchingLatestReleaseOfBase

FailedToFetchLatestReleaseOfBase

HappyCoding

ProjectHasNoReleases ProjectName

UpdateLookingForDependents

UpdateStartTypechecking

UpdateTypecheckingFailure

UpdateTypecheckingSuccess

UpdateIncompleteConstructorSet UpdateOrUpgrade Name (Map ConstructorId Name) (Maybe Int)

UpgradeFailure !ProjectBranchName !ProjectBranchName !FilePath !NameSegment !NameSegment

UpgradeSuccess !NameSegment !NameSegment

MergeFailure !FilePath !MergeSourceAndTarget !ProjectBranchName

MergeSuccess !MergeSourceAndTarget

MergeSuccessFastForward !MergeSourceAndTarget

MergeConflictedAliases !MergeSourceOrTarget !(Defn (Name, Name) (Name, Name))

MergeConflictInvolvingBuiltin !(Defn Name Name)

MergeDefnsInLib !MergeSourceOrTarget

InstalledLibdep !(ProjectAndBranch ProjectName ProjectBranchName) !NameSegment

NoUpgradeInProgress

UseLibInstallNotPull !(ProjectAndBranch ProjectName ProjectBranchName)

PullIntoMissingBranch !(ReadRemoteNamespace RemoteProjectBranch) !(ProjectAndBranch (Maybe ProjectName) ProjectBranchName)

NoMergeInProgress

Output'DebugSynhashTerm !TermReference !Hash !Text

ConflictedDefn !Text !(Defn (Conflicted Name Referent) (Conflicted Name TypeReference))

IncoherentDeclDuringMerge !MergeSourceOrTarget !IncoherentDeclReason

IncoherentDeclDuringUpdate !IncoherentDeclReason

data AmbiguousReset'Argument

Constructors

AmbiguousReset'Hash

AmbiguousReset'Target

data CreatedProjectBranchFrom

What did we create a project branch from?

Loose code

Nothingness (we made an empty branch)

Other branch (in another project)

Parent branch (in this project)

Constructors

CreatedProjectBranchFrom'LooseCode Absolute

CreatedProjectBranchFrom'Nothingness

CreatedProjectBranchFrom'OtherBranch (ProjectAndBranch Project ProjectBranch)

CreatedProjectBranchFrom'ParentBranch ProjectBranchName

data WhichBranchEmpty

A branch was empty. But how do we refer to that branch?

Constructors

WhichBranchEmptyHash ShortCausalHash

WhichBranchEmptyPath ProjectPath

data NumberedOutput

Constructors

ShowDiffNamespace (Either ShortCausalHash ProjectPath) (Either ShortCausalHash ProjectPath) PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterUndo PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterDeleteDefinitions PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterDeleteBranch Absolute PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterModifyBranch Path' Absolute PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterMerge (Either ProjectPath (ProjectAndBranch Project ProjectBranch)) ProjectPath PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterMergePropagate (Either ProjectPath (ProjectAndBranch Project ProjectBranch)) ProjectPath Path' PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterMergePreview (Either ProjectPath (ProjectAndBranch Project ProjectBranch)) ProjectPath PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterPull Path' Absolute PrettyPrintEnv (BranchDiffOutput Symbol Ann)

ShowDiffAfterCreateAuthor NameSegment Path' Absolute PrettyPrintEnv (BranchDiffOutput Symbol Ann)

TestResults TestReportStats PrettyPrintEnv ShowSuccesses ShowFailures (Map TermReferenceId [Text]) (Map TermReferenceId [Text])

Output'Todo !TodoOutput

CantDeleteDefinitions PrettyPrintEnvDecl (Map LabeledDependency (NESet LabeledDependency))
CantDeleteDefinitions ppe couldntDelete becauseTheseStillReferenceThem

CantDeleteNamespace PrettyPrintEnvDecl (Map LabeledDependency (NESet LabeledDependency))
CantDeleteNamespace ppe couldntDelete becauseTheseStillReferenceThem

DeletedDespiteDependents PrettyPrintEnvDecl (Map LabeledDependency (NESet LabeledDependency))
DeletedDespiteDependents ppe deletedThings thingsWhichNowHaveUnnamedReferences

History (Maybe Int) HashLength [(CausalHash, Diff)] HistoryTail
size limit, history , how the history ends

ListProjects [Project]

ListBranches ProjectName [(ProjectBranchName, [(URI, ProjectName, ProjectBranchName)])]

AmbiguousSwitch ProjectName (ProjectAndBranch ProjectName ProjectBranchName)

AmbiguousReset AmbiguousReset'Argument (ProjectAndBranch Project ProjectBranch, Path) (ProjectAndBranch ProjectName ProjectBranchName)

ListNamespaceDependencies PrettyPrintEnv ProjectPath (Map LabeledDependency (Set Name))
List all direct dependencies which don't have any names in the current branch

ShowProjectBranchReflog (Maybe UTCTime) MoreEntriesThanShown [Entry Project ProjectBranch (CausalHash, ShortCausalHash)]

type NumberedArgs = [StructuredArgument]

NB: This only temporarily holds Text. Until all of the inputs are updated to handle StructuredArguments, we need to ensure that the serialization remains unchanged.

type ListDetailed = Bool

data HistoryTail

Constructors

EndOfLog CausalHash

MergeTail CausalHash [CausalHash]

PageEnd CausalHash Int

data TestReportStats

Constructors

CachedTests TotalCount CachedCount

NewlyComputed

data TodoOutput

Constructors

TodoOutput

Fields

defnsInLib :: !Bool

dependentsOfTodo :: !(Set TermReferenceId)

directDependenciesWithoutNames :: !(DefnsF Set TermReference TypeReference)

hashLen :: !Int

incoherentDeclReasons :: !IncoherentDeclReasons

nameConflicts :: !Names

ppe :: !PrettyPrintEnvDecl

todoOutputIsEmpty :: TodoOutput -> Bool

data MoreEntriesThanShown

Constructors

MoreEntriesThanShown

AllEntriesShown

data UndoFailureReason

Constructors

CantUndoPastStart

CantUndoPastMerge

data ShareError

Constructors

ShareErrorDownloadEntities DownloadEntitiesError

ShareErrorGetCausalHashByPath GetCausalHashByPathError

ShareErrorPull PullError

ShareErrorTransport CodeserverTransportError

ShareErrorUploadEntities UploadEntitiesError

ShareExpectedSquashedHead

data UpdateOrUpgrade

Constructors

UOUUpdate

UOUUpgrade

isFailure :: Output -> Bool

isNumberedFailure :: NumberedOutput -> Bool