Unison.Codebase.Transcript.Runner

Description

Execute transcripts.

Documentation

data Error

Constructors

ParseError (ParseErrorBundle Text Void)

RunFailure Text

type Runner = String -> Text -> (FilePath, Codebase IO Symbol Ann) -> IO (Either Error Text)

withRunner

Arguments

:: forall m r. MonadUnliftIO m

=> Bool
Whether to treat this transcript run as a transcript test, which will try to make output deterministic

-> Verbosity

-> UCMVersion

-> FilePath

-> (Runner -> m r)

-> m r