Unison.CommandLine

Documentation

data ParseFailure

Constructors

NoCommand

UnknownCommand String

ExpansionFailure String InputPattern ExpansionFailure

FZFResolveFailure InputPattern FZFResolveFailure

SubParseFailure String InputPattern (Pretty ColorText)

data ExpansionFailure

Constructors

TooManyArguments (NonEmpty Argument)

UnexpectedStructuredArgument StructuredArgument

data FZFResolveFailure

Constructors

NoFZFResolverForArgumentType ParameterDescription

NoFZFOptions

Fields

Text

argument description

allow :: FilePath -> Bool

parseInput

Arguments

:: Codebase IO Symbol Ann

-> ProjectPath
Current location

-> IO (Branch IO)

-> NumberedArgs
Numbered arguments

-> Map String InputPattern
Input Pattern Map

-> [String]
command:arguments

-> IO (Either ParseFailure (Maybe (Arguments, Input)))

prompt :: String

reportParseFailure :: ParseFailure -> Pretty ColorText

TODO: Move this closer to main, but right now it’s shared by ucm and transcripts, so this is the closest we can get without duplicating it.

watchFileSystem :: TQueue Event -> FilePath -> IO (IO ())