Unison.CommandLine.InputPattern

Documentation

data InputPattern

Constructors

InputPattern

Fields

patternName :: String

aliases :: [String]

visibility :: Visibility

args :: [(ArgumentDescription, IsOptional, ArgumentType)]

help :: Pretty ColorText

parse :: Arguments -> Either (Pretty ColorText) Input

Parse the arguments and return either an error message or a command Input.

NB: This function should return Left only on failure. For commands (like $sel:help:InputPattern) that simply produce formatted output, use `pure . Input.CreateMessage`. The failure output should be fully formatted (using wrap, etc.), but shouldn’t include any general error components like a warninng flag or the full help message, and shouldn’t plan for the context it is being output to (e.g., don’t indentN the entire message).

type Argument = Either String StructuredArgument

An argument to a command is either a string provided by the user which needs to be parsed or a numbered argument that doesn’t need to be parsed, as we’ve preserved its representation (although the numbered argument could still be of the wrong type, which should result in an error).

data ArgumentType

Constructors

ArgumentType

Fields

typeName :: String

suggestions :: forall m v a. MonadIO m => String -> Codebase m v a -> AuthenticatedHttpClient -> ProjectPath -> m [Completion]

Generate completion suggestions for this argument type

fzfResolver :: Maybe FZFResolver

If an argument is marked as required, but not provided, the fuzzy finder will be triggered if available.

type ArgumentDescription = Text

Argument description It should fit grammatically into sentences like "I was expecting an argument for the argDesc" e.g. "namespace to merge", "definition to delete", "remote target to push to" etc.

type Arguments = [Argument]

argType :: InputPattern -> Int -> Maybe ArgumentType

data FZFResolver

Constructors

FZFResolver

Fields

getOptions :: OptionFetcher

data IsOptional

Constructors

Required

Optional

ZeroPlus

OnePlus

data Visibility

Constructors

Hidden

Visible

Currently Unused

minArgs :: InputPattern -> Int

maxArgs :: InputPattern -> Maybe Int

unionSuggestions :: forall m v a. MonadIO m => [String -> Codebase m v a -> AuthenticatedHttpClient -> ProjectPath -> m [Completion]] -> String -> Codebase m v a -> AuthenticatedHttpClient -> ProjectPath -> m [Completion]

Union suggestions from all possible completions

suggestionFallbacks :: forall m v a. MonadIO m => [String -> Codebase m v a -> AuthenticatedHttpClient -> ProjectPath -> m [Completion]] -> String -> Codebase m v a -> AuthenticatedHttpClient -> ProjectPath -> m [Completion]

Try the first completer, if it returns no suggestions, try the second, etc.