Unison.CommandLine.InputPatterns

Description

This module defines InputPattern values for every supported input command.

Input commands

add :: InputPattern

aliasMany :: InputPattern

aliasTerm :: InputPattern

aliasType :: InputPattern

api :: InputPattern

authLogin :: InputPattern

back :: InputPattern

branchEmptyInputPattern :: InputPattern

branchInputPattern :: InputPattern

branchRenameInputPattern :: InputPattern

branchesInputPattern :: InputPattern

cd :: InputPattern

clear :: InputPattern

clone :: InputPattern

compileScheme :: InputPattern

createAuthor :: InputPattern

debugClearWatchCache :: InputPattern

debugDoctor :: InputPattern

debugDumpNamespace :: InputPattern

debugDumpNamespaceSimple :: InputPattern

debugFileHashes :: InputPattern

debugFormat :: InputPattern

debugFuzzyOptions :: InputPattern

debugLSPFoldRanges :: InputPattern

debugNameDiff :: InputPattern

debugNumberedArgs :: InputPattern

debugTabCompletion :: InputPattern

debugLspNameCompletion :: InputPattern

debugTerm :: InputPattern

debugTermVerbose :: InputPattern

debugType :: InputPattern

delete :: InputPattern

deleteBranch :: InputPattern

deleteNamespace :: InputPattern

deleteNamespaceForce :: InputPattern

deleteProject :: InputPattern

deleteTerm :: InputPattern

deleteTermVerbose :: InputPattern

deleteType :: InputPattern

deleteTypeVerbose :: InputPattern

deleteVerbose :: InputPattern

dependencies :: InputPattern

dependents :: InputPattern

diffNamespace :: InputPattern

display :: InputPattern

displayTo :: InputPattern

docToMarkdown :: InputPattern

docs :: InputPattern

docsToHtml :: InputPattern

edit :: InputPattern

editNamespace :: InputPattern

execute :: InputPattern

find :: InputPattern

findAll :: InputPattern

findGlobal :: InputPattern

findIn :: InputPattern

findInAll :: InputPattern

findShallow :: InputPattern

findVerbose :: InputPattern

findVerboseAll :: InputPattern

forkLocal :: InputPattern

help :: InputPattern

helpTopics :: InputPattern

history :: InputPattern

ioTest :: InputPattern

ioTestAll :: InputPattern

libInstallInputPattern :: InputPattern

load :: InputPattern

makeStandalone :: InputPattern

mergeBuiltins :: InputPattern

mergeCommitInputPattern :: InputPattern

mergeIOBuiltins :: InputPattern

mergeInputPattern :: InputPattern

mergeOldInputPattern :: InputPattern

mergeOldPreviewInputPattern :: InputPattern

mergeOldSquashInputPattern :: InputPattern

moveAll :: InputPattern

names :: IsGlobal -> InputPattern

namespaceDependencies :: InputPattern

previewAdd :: InputPattern

previewUpdate :: InputPattern

printVersion :: InputPattern

projectCreate :: InputPattern

projectCreateEmptyInputPattern :: InputPattern

projectRenameInputPattern :: InputPattern

projectSwitch :: InputPattern

projectsInputPattern :: InputPattern

pull :: InputPattern

pullWithoutHistory :: InputPattern

push :: InputPattern

pushCreate :: InputPattern

pushExhaustive :: InputPattern

pushForce :: InputPattern

quit :: InputPattern

releaseDraft :: InputPattern

renameBranch :: InputPattern

renameTerm :: InputPattern

renameType :: InputPattern

reset :: InputPattern

runScheme :: InputPattern

saveExecuteResult :: InputPattern

sfind :: InputPattern

sfindReplace :: InputPattern

textfind :: Bool -> InputPattern

test :: InputPattern

testAll :: InputPattern

todo :: InputPattern

ui :: InputPattern

undo :: InputPattern

up :: InputPattern

update :: InputPattern

updateBuiltins :: InputPattern

updateOld :: InputPattern

updateOldNoPatch :: InputPattern

upgrade :: InputPattern

upgradeCommitInputPattern :: InputPattern

view :: InputPattern

viewGlobal :: InputPattern

deprecatedViewRootReflog :: InputPattern

branchReflog :: InputPattern

projectReflog :: InputPattern

globalReflog :: InputPattern

Misc

formatStructuredArgument :: Maybe Int -> StructuredArgument -> Text

helpFor :: InputPattern -> Pretty ColorText

makeExample' :: InputPattern -> Pretty ColorText

makeExample :: InputPattern -> [Pretty ColorText] -> Pretty ColorText

makeExampleEOS :: InputPattern -> [Pretty ColorText] -> Pretty ColorText

makeExampleNoBackticks :: InputPattern -> [Pretty ColorText] -> Pretty ColorText

patternMap :: Map String InputPattern

A map of all command patterns by pattern name or alias.

patternName :: InputPattern -> Pretty ColorText

showPatternHelp :: InputPattern -> Pretty ColorText

unifyArgument :: Argument -> String

Converts an arbitrary argument to a String. This is for cases where the command should accept a structured argument of some type, but currently wants a String.

validInputs :: [InputPattern]