Unison.CommandLine.OutputMessages

Documentation

reportBugURL :: Pretty

type Pretty = Pretty ColorText

shortenDirectory :: FilePath -> IO FilePath

renderFileName :: FilePath -> IO Pretty

notifyNumbered :: NumberedOutput -> (Pretty, NumberedArgs)

undoTip :: Pretty ColorText

notifyUser :: FilePath -> Output -> IO Pretty

prettyShareError :: ShareError -> Pretty

prettyDownloadEntitiesError :: DownloadEntitiesError -> Pretty

prettyGetCausalHashByPathError :: GetCausalHashByPathError -> Pretty

prettyPullError :: PullError -> Pretty

prettyUploadEntitiesError :: UploadEntitiesError -> Pretty

prettyEntityValidationFailure :: EntityValidationError -> Pretty

prettyTransportError :: CodeserverTransportError -> Pretty

prettyEntityType :: EntityType -> Pretty

invalidRepoInfo :: Text -> RepoInfo -> Pretty

hashMismatchFromShare :: Hash32 -> Hash32 -> Pretty

pushPublicNote :: InputPattern -> Text -> [Text] -> Pretty

needDependencies :: NeedDependencies Hash32 -> Pretty

noReadPermissionForPath :: Path -> Pretty

noReadPermissionForRepo :: RepoInfo -> Pretty

noWritePermissionForPath :: Path -> Pretty

noWritePermissionForRepo :: RepoInfo -> Pretty

notFastForward :: Path -> Pretty

shareProjectNotFound :: Text -> Pretty

shareUserNotFound :: RepoInfo -> Pretty

formatMissingStuff :: (Show tm, Show typ) => [(HashQualified Name, tm)] -> [(HashQualified Name, typ)] -> Pretty

displayOutputRewrittenFile :: Var v => FilePath -> [v] -> IO Pretty

displayDefinitions' :: Var v => Ord a1 => PrettyPrintEnvDecl -> Map Reference (DisplayObject () (Decl v a1)) -> Map Reference (DisplayObject (Type v a1) (Term v a1)) -> Pretty

displayRendered :: Maybe FilePath -> Pretty -> IO Pretty

displayTestResults :: Bool -> [(HashQualified Name, [Text])] -> [(HashQualified Name, [Text])] -> Pretty

unsafePrettyTermResultSig' :: Var v => PrettyPrintEnv -> TermResult' v a -> Pretty

renderNameConflicts :: Int -> Names -> Numbered Pretty

type Numbered = State (Int, Seq StructuredArgument)

addNumberedArg :: StructuredArgument -> Numbered Int

formatNum :: Int -> Pretty

runNumbered :: Numbered a -> (a, NumberedArgs)

handleTodoOutput :: TodoOutput -> Numbered Pretty

listOfDefinitions :: Var v => FindScope -> PrettyPrintEnv -> ListDetailed -> [SearchResult' v a] -> IO Pretty

listOfNames :: Int -> [(Reference, [HashQualified Name])] -> [(Referent, [HashQualified Name])] -> IO Pretty

data ShowNumbers

Constructors

ShowNumbers

HideNumbers

showDiffNamespace :: forall v. Var v => ShowNumbers -> PrettyPrintEnv -> AbsBranchId -> AbsBranchId -> BranchDiffOutput v Ann -> (Pretty, NumberedArgs)

ppe is just for rendering type signatures `oldPath, newPath :: Path.Absolute` are just for producing fully-qualified numbered args

noResults :: FindScope -> Pretty

listOfDefinitions' :: Var v => FindScope -> PrettyPrintEnv -> ListDetailed -> [SearchResult' v a] -> Pretty

watchPrinter :: Var v => Text -> PrettyPrintEnv -> Ann -> WatchKind -> Term v () -> IsCacheHit -> Pretty

filestatusTip :: Pretty

prettyDiff :: Diff -> Pretty

numberedArgsForEndangerments :: PrettyPrintEnvDecl -> Map LabeledDependency (NESet LabeledDependency) -> NumberedArgs

Get the list of numbered args corresponding to an endangerment map, which is used by a few outputs. See endangeredDependentsTable.

endangeredDependentsTable :: PrettyPrintEnvDecl -> Map LabeledDependency (NESet LabeledDependency) -> Pretty ColorText

Format and render all dependents which are endangered by references going extinct.

listFind :: Bool -> Maybe Pretty -> [HashQualified Name] -> Pretty

listDependentsOrDependencies :: PrettyPrintEnv -> Text -> Text -> Set LabeledDependency -> [HashQualified Name] -> [HashQualified Name] -> Pretty

displayProjectBranchReflogEntries :: Maybe UTCTime -> MoreEntriesThanShown -> [Entry Project ProjectBranch (CausalHash, ShortCausalHash)] -> (Pretty, NumberedArgs)