Unison.LSP.Diagnostics

Documentation

annToRange :: Ann -> Maybe Range

uToLspPos :: Pos -> Position

uToLspRange :: Range -> Range

reportDiagnostics

Arguments

:: Foldable f

=> Uri

-> Maybe FileVersion

-> f Diagnostic
Note, it's important to still send an empty list of diagnostics if there aren't any because it clears existing diagnostics in the editor

-> Lsp ()

mkDiagnostic :: Uri -> Range -> DiagnosticSeverity -> Text -> [(Text, Range)] -> Diagnostic