Unison.LSP.Diagnostics

Documentation

reportDiagnostics

Arguments

:: Foldable f

=> Uri

-> Maybe FileVersion

-> f Diagnostic
Note, it's important to still send an empty list of diagnostics if there aren't any because it clears existing diagnostics in the editor

-> Lsp ()

mkDiagnostic :: Uri -> Range -> DiagnosticSeverity -> [DiagnosticTag] -> Text -> [(Text, Range)] -> Diagnostic

data DiagnosticSeverity

Constructors

DiagnosticSeverity_Error

DiagnosticSeverity_Warning

DiagnosticSeverity_Information

DiagnosticSeverity_Hint