Unison.LSP.Types

Documentation

newtype Lsp a

A custom LSP monad wrapper so we can provide our own environment.

Constructors

Lsp

Fields

runLspM :: ReaderT Env (LspM Config) a

logInfo :: Text -> Lsp ()

Log an info message to the client's LSP log.

logError :: Text -> Lsp ()

Log an error message to the client's LSP log, this will be shown to the user in most LSP implementations.

data Env

Environment for the Lsp monad.

Constructors

Env

Fields

lspContext :: LanguageContextEnv Config

codebase :: Codebase IO Symbol Ann

parseNamesCache :: IO NamesWithHistory

ppedCache :: IO PrettyPrintEnvDecl

currentPathCache :: IO Absolute

vfsVar :: MVar VFS

runtime :: Runtime Symbol

checkedFilesVar :: TVar (Map Uri FileAnalysis)

dirtyFilesVar :: TVar (Set Uri)

cancellationMapVar :: TVar (Map SomeLspId (IO ()))

completionsVar :: TVar CompletionTree

scope :: Scope

newtype CompletionTree

A suffix tree over path segments of name completions. see namesToCompletionTree for more on how this is built and the invariants it should have.

Constructors

CompletionTree

Fields

unCompletionTree :: Cofree (Map NameSegment) (Set (Name, LabeledDependency))

type FileVersion = Int32

A monotonically increasing file version tracked by the lsp client.

type LexedSource = (Text, [Token Lexeme])

data FileAnalysis

Constructors

FileAnalysis

Fields

fileUri :: Uri

fileVersion :: FileVersion

lexedSource :: LexedSource

tokenMap :: IntervalMap Position Lexeme

parsedFile :: Maybe (UnisonFile Symbol Ann)

typecheckedFile :: Maybe (TypecheckedUnisonFile Symbol Ann)

notes :: Seq (Note Symbol Ann)

diagnostics :: IntervalMap Position [Diagnostic]

codeActions :: IntervalMap Position [CodeAction]

fileSummary :: Maybe FileSummary

data FileSummary

A file that parses might not always type-check, but often we just want to get as much information as we have available. This provides a type where we can summarize the information available in a Unison file.

If the file typechecked then all the Ref Ids and types will be filled in, otherwise they will be Nothing.

Constructors

FileSummary

Fields

dataDeclsBySymbol :: Map Symbol (Id, DataDeclaration Symbol Ann)

dataDeclsByReference :: Map Id (Map Symbol (DataDeclaration Symbol Ann))

effectDeclsBySymbol :: Map Symbol (Id, EffectDeclaration Symbol Ann)

effectDeclsByReference :: Map Id (Map Symbol (EffectDeclaration Symbol Ann))

termsBySymbol :: Map Symbol (Maybe Id, Term Symbol Ann, Maybe (Type Symbol Ann))

termsByReference :: Map (Maybe Id) (Map Symbol (Term Symbol Ann, Maybe (Type Symbol Ann)))

testWatchSummary :: [(Maybe Symbol, Maybe Id, Term Symbol Ann, Maybe (Type Symbol Ann))]

exprWatchSummary :: [(Maybe Symbol, Maybe Id, Term Symbol Ann, Maybe (Type Symbol Ann))]

fileNames :: Names

getCurrentPath :: Lsp Absolute

getCodebaseCompletions :: Lsp CompletionTree

globalPPED :: Lsp PrettyPrintEnvDecl

getParseNames :: Lsp NamesWithHistory

data Config

Constructors

Config

Fields

maxCompletions :: Maybe Int

defaultLSPConfig :: Config

lspBackend :: Backend IO a -> Lsp (Either BackendError a)

Lift a backend computation into the Lsp monad.

sendNotification :: forall (m :: Method 'FromServer 'Notification). Message m ~ NotificationMessage m => NotificationMessage m -> Lsp ()

data RangedCodeAction

Constructors

RangedCodeAction

Fields

_codeActionRanges :: [Range]

_codeAction :: CodeAction

rangedCodeAction :: Text -> [Diagnostic] -> [Range] -> RangedCodeAction

includeEdits :: Uri -> Text -> [Range] -> RangedCodeAction -> RangedCodeAction

Provided ranges must not intersect.

getConfig :: Lsp Config

setConfig :: Config -> Lsp ()