Unison.LSP.Types

Documentation

newtype Lsp a

A custom LSP monad wrapper so we can provide our own environment.

Constructors

Lsp

Fields

runLspM :: ReaderT Env (LspM Config) a

logInfo :: Text -> Lsp ()

Log an info message to the client's LSP log.

logError :: Text -> Lsp ()

Log an error message to the client's LSP log, this will be shown to the user in most LSP implementations.

data Env

Environment for the Lsp monad.

Constructors

Env

Fields

lspContext :: LanguageContextEnv Config

codebase :: Codebase IO Symbol Ann

currentNamesCache :: IO Names

ppedCache :: IO PrettyPrintEnvDecl

nameSearchCache :: IO (NameSearch Transaction)

currentProjectPathCache :: IO ProjectPath

vfsVar :: MVar VFS

runtime :: Runtime Symbol

checkedFilesVar :: TVar (Map Uri (TMVar FileAnalysis))

dirtyFilesVar :: TVar (Set Uri)

cancellationMapVar :: TVar (Map (Int32 |? Text) (IO ()))

completionsVar :: TMVar CompletionTree

scope :: Scope

newtype CompletionTree

A suffix tree over path segments of name completions. see namesToCompletionTree for more on how this is built and the invariants it should have.

Constructors

CompletionTree

Fields

unCompletionTree :: Cofree (Map NameSegment) (Set (Name, LabeledDependency))

type FileVersion = Int32

A monotonically increasing file version tracked by the lsp client.

type LexedSource = (Text, [Token Lexeme])

data TypeSignatureHint

Constructors

TypeSignatureHint

Fields

name :: Name

referent :: Referent

bindingLocation :: Range

signature :: Type Symbol Ann

data FileAnalysis

Constructors

FileAnalysis

Fields

fileUri :: Uri

fileVersion :: FileVersion

lexedSource :: LexedSource

tokenMap :: IntervalMap Position Lexeme

parsedFile :: Maybe (UnisonFile Symbol Ann)

typecheckedFile :: Maybe (TypecheckedUnisonFile Symbol Ann)

notes :: Seq (Note Symbol Ann)

diagnostics :: IntervalMap Position [Diagnostic]

codeActions :: IntervalMap Position [CodeAction]

typeSignatureHints :: Map Symbol TypeSignatureHint

fileSummary :: Maybe FileSummary

getCurrentProjectPath :: Lsp ProjectPath

getCodebaseCompletions :: Lsp CompletionTree

currentPPED :: Lsp PrettyPrintEnvDecl

getNameSearch :: Lsp (NameSearch Transaction)

getCurrentNames :: Lsp Names

data Config

Constructors

Config

Fields

formattingWidth :: Int

maxCompletions :: Maybe Int

defaultLSPConfig :: Config

lspBackend :: Backend IO a -> Lsp (Either BackendError a)

Lift a backend computation into the Lsp monad.

sendNotification :: forall (m :: Method 'ServerToClient 'Notification). TMessage m ~ TNotificationMessage m => TNotificationMessage m -> Lsp ()

data RangedCodeAction

Constructors

RangedCodeAction

Fields

_codeActionRanges :: [Range]

_codeAction :: CodeAction

rangedCodeAction :: Text -> [Diagnostic] -> [Range] -> RangedCodeAction

includeEdits :: Uri -> Text -> [Range] -> RangedCodeAction -> RangedCodeAction

Provided ranges must not intersect.

getConfig :: Lsp Config

setConfig :: Config -> Lsp ()