Unison.Share.Sync.Types

Description

Types used by the UCM client during sync.

Documentation

data CodeserverTransportError

Generic Codeserver transport errors

Constructors

DecodeFailure Text Response

PermissionDenied Text

RateLimitExceeded

Timeout

Unauthenticated BaseUrl

UnexpectedResponse Response

UnreachableCodeserver BaseUrl

data GetCausalHashByPathError

An error occurred when getting causal hash by path.

Constructors

GetCausalHashByPathErrorNoReadPermission Path
The user does not have permission to read this path.

GetCausalHashByPathErrorInvalidRepoInfo Text RepoInfo
The repo info was invalid. (err, repoInfo)

GetCausalHashByPathErrorUserNotFound RepoInfo
The user was not found.

data PullError

An error occurred while pulling code from Unison Share.

Constructors

PullError'DownloadEntities DownloadEntitiesError

PullError'GetCausalHash GetCausalHashByPathError

PullError'NoHistoryAtPath Path

data SyncError e

Constructors

TransportError CodeserverTransportError

SyncError e